Module harmony_rust_sdk::api

Harmony protocol API.

Modules

auth

Auth service API.

batch

Batch service API.

bots

Bots service API.

chat

Chat service API.

emote

Emote service API.

exports

Some crates re-exported for user convenience.

harmonytypes

Common types used in other services.

mediaproxy

Media proxy service API.

profile

Profile service API.

rest

Rest API common types.

sync

Sync service API.

voice

Voice service API.

Structs

Hmc

A HMC.

HomeserverIdentifier

A homeserver identifier containg a domain and a port.

Enums

HmcFromStrError

Errors that can occur when converting a possibly non-HMC string to a HMC URL.

HmcParseError

Errors that can occur when converting a possibly non-HMC URL to a HMC URL.

HomeserverIdParseError

Error thrown when parsing a homeserver ID from a string.

Traits

Endpoint

A trait to facilitate easy request execution.