Module harmony_rust_sdk::api::auth::v1

Modules

auth_service_client

Generated client implementations.

auth_service_server

Generated server implementations.

auth_step

Nested message and enum types in AuthStep.

next_step_request

Nested message and enum types in NextStepRequest.

Structs

ArchivedAuthStep

An archived AuthStep

ArchivedBeginAuthRequest

An archived BeginAuthRequest

ArchivedBeginAuthResponse

An archived BeginAuthResponse

ArchivedCheckLoggedInRequest

An archived CheckLoggedInRequest

ArchivedCheckLoggedInResponse

An archived CheckLoggedInResponse

ArchivedFederateRequest

An archived FederateRequest

ArchivedFederateResponse

An archived FederateResponse

ArchivedKeyRequest

An archived KeyRequest

ArchivedKeyResponse

An archived KeyResponse

ArchivedLoginFederatedRequest

An archived LoginFederatedRequest

ArchivedLoginFederatedResponse

An archived LoginFederatedResponse

ArchivedNextStepRequest

An archived NextStepRequest

ArchivedNextStepResponse

An archived NextStepResponse

ArchivedSession

An archived Session

ArchivedStepBackRequest

An archived StepBackRequest

ArchivedStepBackResponse

An archived StepBackResponse

ArchivedStreamStepsRequest

An archived StreamStepsRequest

ArchivedStreamStepsResponse

An archived StreamStepsResponse

ArchivedTokenData

An archived TokenData

AuthStep

AuthStep A step in the authentication process Contains a variety of different types of views It is recommended to have a fallback_url specified For non-trivial authentication procedures (such as captchas)

AuthStepResolver

The resolver for an archived AuthStep

BeginAuthRequest

Used in BeginAuth endpoint.

BeginAuthRequestResolver

The resolver for an archived BeginAuthRequest

BeginAuthResponse

BeginAuthResponse The return type of BeginAuth, containing the auth_id that will be used for the authentication section

BeginAuthResponseResolver

The resolver for an archived BeginAuthResponse

CheckLoggedInRequest

Used in CheckLoggedIn endpoint.

CheckLoggedInRequestResolver

The resolver for an archived CheckLoggedInRequest

CheckLoggedInResponse

Used in CheckLoggedIn endpoint.

CheckLoggedInResponseResolver

The resolver for an archived CheckLoggedInResponse

FederateRequest

The request to federate with a foreign server.

FederateRequestResolver

The resolver for an archived FederateRequest

FederateResponse

The reply to a successful federation request, containing the token you need to present to the foreign server.

FederateResponseResolver

The resolver for an archived FederateResponse

KeyRequest

Used in Key endpoint.

KeyRequestResolver

The resolver for an archived KeyRequest

KeyResponse

Contains a key’s bytes.

KeyResponseResolver

The resolver for an archived KeyResponse

LoginFederatedRequest

Log into a foreignserver using a token from your homeserver, obtained through a FederateRequest

LoginFederatedRequestResolver

The resolver for an archived LoginFederatedRequest

LoginFederatedResponse

Used in LoginFederated endpoint.

LoginFederatedResponseResolver

The resolver for an archived LoginFederatedResponse

NextStepRequest

NextStepRequest contains the client’s response to the server’s challenge This needs to be called first with no arguments to receive the first step

NextStepRequestResolver

The resolver for an archived NextStepRequest

NextStepResponse

Used in NextStep endpoint.

NextStepResponseResolver

The resolver for an archived NextStepResponse

Session

Session Session contains the information for a new session; the user_id you logged in as and the session_token which should be passed to authorisation

SessionResolver

The resolver for an archived Session

StepBackRequest

StepBackRequest A request to go back 1 step

StepBackRequestResolver

The resolver for an archived StepBackRequest

StepBackResponse

Used in StepBack endpoint.

StepBackResponseResolver

The resolver for an archived StepBackResponse

StreamStepsRequest

StreamStepsRequest Required to be initiated by all authenticating clients Allows the server to send steps

StreamStepsRequestResolver

The resolver for an archived StreamStepsRequest

StreamStepsResponse

Used in StreamSteps endpoint.

StreamStepsResponseResolver

The resolver for an archived StreamStepsResponse

TokenData

Information sent by a client’s homeserver, in a harmonytypes.v1.Token. It will be sent to a foreignserver by the client.

TokenDataResolver

The resolver for an archived TokenData