Trait harmony_rust_sdk::api::auth::v1::auth_service_server::AuthService

pub trait AuthService: Sync + Send + 'static {

    fn federate(
        &self, 
        request: HrpcRequest<FederateRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<FederateResponse>>>;

    fn login_federated(
        &self, 
        request: HrpcRequest<LoginFederatedRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<LoginFederatedResponse>>>;

    fn key(
        &self, 
        request: HrpcRequest<KeyRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<KeyResponse>>>;

    fn begin_auth(
        &self, 
        request: HrpcRequest<BeginAuthRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<BeginAuthResponse>>>;

    fn next_step(
        &self, 
        request: HrpcRequest<NextStepRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<NextStepResponse>>>;

    fn step_back(
        &self, 
        request: HrpcRequest<StepBackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<StepBackResponse>>>;

    fn stream_steps(
        &self, 
        request: HrpcRequest<()>, 
        socket: Socket<StreamStepsResponse, StreamStepsRequest>
    ) -> BoxFuture<'_, ServerResult<()>>;

    fn check_logged_in(
        &self, 
        request: HrpcRequest<CheckLoggedInRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CheckLoggedInResponse>>>;

    fn federate_middleware(&self) -> Option<HrpcLayer> { ... }

    fn login_federated_middleware(&self) -> Option<HrpcLayer> { ... }

    fn key_middleware(&self) -> Option<HrpcLayer> { ... }

    fn begin_auth_middleware(&self) -> Option<HrpcLayer> { ... }

    fn next_step_middleware(&self) -> Option<HrpcLayer> { ... }

    fn step_back_middleware(&self) -> Option<HrpcLayer> { ... }

    fn stream_steps_middleware(&self) -> Option<HrpcLayer> { ... }

    fn check_logged_in_middleware(&self) -> Option<HrpcLayer> { ... }

}

Generated trait containing hRPC methods that should be implemented for use with AuthServiceServer.

Required methods

fn federate(
&self,
request: HrpcRequest<FederateRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<FederateResponse>>>

Federate with a foreignserver, obtaining a token you can use to call LoginFederated on it

fn login_federated(
&self,
request: HrpcRequest<LoginFederatedRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<LoginFederatedResponse>>>

Present a token to a foreignserver from a Federate call on your homeserver in order to login

fn key(
&self,
request: HrpcRequest<KeyRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<KeyResponse>>>

Returns the public key of this server

fn begin_auth(
&self,
request: HrpcRequest<BeginAuthRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<BeginAuthResponse>>>

Begins an authentication session

fn next_step(
&self,
request: HrpcRequest<NextStepRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<NextStepResponse>>>

Goes to the next step of the authentication session, possibly presenting user input

fn step_back(
&self,
request: HrpcRequest<StepBackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<StepBackResponse>>>

Goes to the previous step of the authentication session if possible

fn stream_steps(
&self,
request: HrpcRequest<()>,
socket: Socket<StreamStepsResponse, StreamStepsRequest>
) -> BoxFuture<'_, ServerResult<()>>

Consume the steps of an authentication session as a stream

fn check_logged_in(
&self,
request: HrpcRequest<CheckLoggedInRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CheckLoggedInResponse>>>

Check whether or not you’re logged in and the session is valid

Provided methods

fn federate_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn login_federated_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn key_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn begin_auth_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn next_step_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn step_back_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn stream_steps_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn check_logged_in_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.