Module harmony_rust_sdk::api::auth::v1::auth_step

Nested message and enum types in AuthStep.

Modules

form

Nested message and enum types in Form.

Structs

ArchivedChoice

An archived Choice

ArchivedForm

An archived Form

ArchivedWaiting

An archived Waiting

Choice

Choice A step which allows the user to choose from a range of options Allows you to show a heading by specifying title

ChoiceResolver

The resolver for an archived Choice

Form

Form A step which requires the user to input information Allows you to show a heading by specifying title

FormResolver

The resolver for an archived Form

Waiting

Waiting A step which requires the user to perform an external action The title and description should explain to the user what they should do to complete this step

WaitingResolver

The resolver for an archived Waiting

Enums

ArchivedStep

An archived Step

Step

step: the current step

StepResolver

The resolver for an archived Step