Module harmony_rust_sdk::api::auth::v1::next_step_request

Nested message and enum types in NextStepRequest.

Modules

form_fields
Nested message and enum types in FormFields.

Structs

Choice
A simple choice string indicating which option the user chose

Form
An array of form fields, in the same order they came in from the server

FormFields
Form fields can either be bytes, string, or int64.

Enums

Step
step: the user’s response to a step