Trait harmony_rust_sdk::api::batch::v1::batch_service_server::BatchService

pub trait BatchService: Sync + Send + 'static {
    fn batch(
        &self, 
        request: HrpcRequest<BatchRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<BatchResponse>>>;

    fn batch_same(
        &self, 
        request: HrpcRequest<BatchSameRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<BatchSameResponse>>>;

    fn batch_middleware(&self) -> Option<HrpcLayer> { ... }

    fn batch_same_middleware(&self) -> Option<HrpcLayer> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with BatchServiceServer.

Required methods

fn batch(
&self,
request: HrpcRequest<BatchRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<BatchResponse>>>

Batch requests. Does not support batching stream requests. Batched requests should be verified and an error should be thrown if they are invalid.

fn batch_same(
&self,
request: HrpcRequest<BatchSameRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<BatchSameResponse>>>

BatchSame allows batching for requests using the same endpoint. This allows for additional network optimizations since the endpoint doesn’t have to be sent for every request.

Provided methods

fn batch_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn batch_same_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.