Module harmony_rust_sdk::api::chat::v1

Modules

action

Nested message and enum types in Action.

action_payload

Nested message and enum types in ActionPayload.

all_permissions

All chat permissions.

chat_service_client

Generated client implementations.

chat_service_server

Generated server implementations.

content

Nested message and enum types in Content.

embed

Nested message and enum types in Embed.

format

Nested message and enum types in Format.

get_channel_messages_request

Nested message and enum types in GetChannelMessagesRequest.

guild_kind

Nested message and enum types in GuildKind.

overrides

Nested message and enum types in Overrides.

stream_event

Nested message and enum types in StreamEvent.

stream_events_request

Nested message and enum types in StreamEventsRequest.

stream_events_response

Nested message and enum types in StreamEventsResponse.

Structs

Action

Actions are interactive elements that can exist within an embed.

ActionPayload

The payload sent to the bot when an action is triggered.

ActionPayloadResolver

The resolver for an archived ActionPayload

ActionResolver

The resolver for an archived Action

AddGuildRoleRequest

Used in the AddGuildRole endpoint.

AddGuildRoleRequestResolver

The resolver for an archived AddGuildRoleRequest

AddGuildRoleResponse

Used in the AddGuildRole endpoint.

AddGuildRoleResponseResolver

The resolver for an archived AddGuildRoleResponse

AddReactionRequest

Used in AddReaction endpoint.

AddReactionRequestResolver

The resolver for an archived AddReactionRequest

AddReactionResponse

Used in AddReaction endpoint.

AddReactionResponseResolver

The resolver for an archived AddReactionResponse

ArchivedAction

An archived Action

ArchivedActionPayload

An archived ActionPayload

ArchivedAddGuildRoleRequest

An archived AddGuildRoleRequest

ArchivedAddGuildRoleResponse

An archived AddGuildRoleResponse

ArchivedAddReactionRequest

An archived AddReactionRequest

ArchivedAddReactionResponse

An archived AddReactionResponse

ArchivedAttachment

An archived Attachment

ArchivedBanUserRequest

An archived BanUserRequest

ArchivedBanUserResponse

An archived BanUserResponse

ArchivedChannel

An archived Channel

ArchivedChannelWithId

An archived ChannelWithId

ArchivedContent

An archived Content

ArchivedCreateChannelRequest

An archived CreateChannelRequest

ArchivedCreateChannelResponse

An archived CreateChannelResponse

ArchivedCreateDirectMessageRequest

An archived CreateDirectMessageRequest

ArchivedCreateDirectMessageResponse

An archived CreateDirectMessageResponse

ArchivedCreateGuildRequest

An archived CreateGuildRequest

ArchivedCreateGuildResponse

An archived CreateGuildResponse

ArchivedCreateInviteRequest

An archived CreateInviteRequest

ArchivedCreateInviteResponse

An archived CreateInviteResponse

ArchivedCreateRoomRequest

An archived CreateRoomRequest

ArchivedCreateRoomResponse

An archived CreateRoomResponse

ArchivedDeleteChannelRequest

An archived DeleteChannelRequest

ArchivedDeleteChannelResponse

An archived DeleteChannelResponse

ArchivedDeleteGuildRequest

An archived DeleteGuildRequest

ArchivedDeleteGuildResponse

An archived DeleteGuildResponse

ArchivedDeleteGuildRoleRequest

An archived DeleteGuildRoleRequest

ArchivedDeleteGuildRoleResponse

An archived DeleteGuildRoleResponse

ArchivedDeleteInviteRequest

An archived DeleteInviteRequest

ArchivedDeleteInviteResponse

An archived DeleteInviteResponse

ArchivedDeleteMessageRequest

An archived DeleteMessageRequest

ArchivedDeleteMessageResponse

An archived DeleteMessageResponse

ArchivedEmbed

An archived Embed

ArchivedFormat

An archived Format

ArchivedFormattedText

An archived FormattedText

ArchivedGetBannedUsersRequest

An archived GetBannedUsersRequest

ArchivedGetBannedUsersResponse

An archived GetBannedUsersResponse

ArchivedGetChannelMessagesRequest

An archived GetChannelMessagesRequest

ArchivedGetChannelMessagesResponse

An archived GetChannelMessagesResponse

ArchivedGetGuildChannelsRequest

An archived GetGuildChannelsRequest

ArchivedGetGuildChannelsResponse

An archived GetGuildChannelsResponse

ArchivedGetGuildInvitesRequest

An archived GetGuildInvitesRequest

ArchivedGetGuildInvitesResponse

An archived GetGuildInvitesResponse

ArchivedGetGuildListRequest

An archived GetGuildListRequest

ArchivedGetGuildListResponse

An archived GetGuildListResponse

ArchivedGetGuildMembersRequest

An archived GetGuildMembersRequest

ArchivedGetGuildMembersResponse

An archived GetGuildMembersResponse

ArchivedGetGuildRequest

An archived GetGuildRequest

ArchivedGetGuildResponse

An archived GetGuildResponse

ArchivedGetGuildRolesRequest

An archived GetGuildRolesRequest

ArchivedGetGuildRolesResponse

An archived GetGuildRolesResponse

ArchivedGetMessageRequest

An archived GetMessageRequest

ArchivedGetMessageResponse

An archived GetMessageResponse

ArchivedGetPendingInvitesRequest

An archived GetPendingInvitesRequest

ArchivedGetPendingInvitesResponse

An archived GetPendingInvitesResponse

ArchivedGetPermissionsRequest

An archived GetPermissionsRequest

ArchivedGetPermissionsResponse

An archived GetPermissionsResponse

ArchivedGetPinnedMessagesRequest

An archived GetPinnedMessagesRequest

ArchivedGetPinnedMessagesResponse

An archived GetPinnedMessagesResponse

ArchivedGetUserRolesRequest

An archived GetUserRolesRequest

ArchivedGetUserRolesResponse

An archived GetUserRolesResponse

ArchivedGiveUpOwnershipRequest

An archived GiveUpOwnershipRequest

ArchivedGiveUpOwnershipResponse

An archived GiveUpOwnershipResponse

ArchivedGrantOwnershipRequest

An archived GrantOwnershipRequest

ArchivedGrantOwnershipResponse

An archived GrantOwnershipResponse

ArchivedGuild

An archived Guild

ArchivedGuildKind

An archived GuildKind

ArchivedGuildListEntry

An archived GuildListEntry

ArchivedGuildWithId

An archived GuildWithId

ArchivedIgnorePendingInviteRequest

An archived IgnorePendingInviteRequest

ArchivedIgnorePendingInviteResponse

An archived IgnorePendingInviteResponse

ArchivedInvite

An archived Invite

ArchivedInviteUserToGuildRequest

An archived InviteUserToGuildRequest

ArchivedInviteUserToGuildResponse

An archived InviteUserToGuildResponse

ArchivedInviteWithId

An archived InviteWithId

ArchivedJoinGuildRequest

An archived JoinGuildRequest

ArchivedJoinGuildResponse

An archived JoinGuildResponse

ArchivedKickUserRequest

An archived KickUserRequest

ArchivedKickUserResponse

An archived KickUserResponse

ArchivedLeaveGuildRequest

An archived LeaveGuildRequest

ArchivedLeaveGuildResponse

An archived LeaveGuildResponse

ArchivedManageUserRolesRequest

An archived ManageUserRolesRequest

ArchivedManageUserRolesResponse

An archived ManageUserRolesResponse

ArchivedMessage

An archived Message

ArchivedMessageWithId

An archived MessageWithId

ArchivedMinithumbnail

An archived Minithumbnail

ArchivedModifyGuildRoleRequest

An archived ModifyGuildRoleRequest

ArchivedModifyGuildRoleResponse

An archived ModifyGuildRoleResponse

ArchivedMoveRoleRequest

An archived MoveRoleRequest

ArchivedMoveRoleResponse

An archived MoveRoleResponse

ArchivedOverrides

An archived Overrides

ArchivedPendingInvite

An archived PendingInvite

ArchivedPermission

An archived Permission

ArchivedPhoto

An archived Photo

ArchivedPinMessageRequest

An archived PinMessageRequest

ArchivedPinMessageResponse

An archived PinMessageResponse

ArchivedPreviewGuildRequest

An archived PreviewGuildRequest

ArchivedPreviewGuildResponse

An archived PreviewGuildResponse

ArchivedQueryHasPermissionRequest

An archived QueryHasPermissionRequest

ArchivedQueryHasPermissionResponse

An archived QueryHasPermissionResponse

ArchivedReaction

An archived Reaction

ArchivedRejectPendingInviteRequest

An archived RejectPendingInviteRequest

ArchivedRejectPendingInviteResponse

An archived RejectPendingInviteResponse

ArchivedRemoveReactionRequest

An archived RemoveReactionRequest

ArchivedRemoveReactionResponse

An archived RemoveReactionResponse

ArchivedRole

An archived Role

ArchivedRoleWithId

An archived RoleWithId

ArchivedSendMessageRequest

An archived SendMessageRequest

ArchivedSendMessageResponse

An archived SendMessageResponse

ArchivedSetPermissionsRequest

An archived SetPermissionsRequest

ArchivedSetPermissionsResponse

An archived SetPermissionsResponse

ArchivedStreamEvent

An archived StreamEvent

ArchivedStreamEventsRequest

An archived StreamEventsRequest

ArchivedStreamEventsResponse

An archived StreamEventsResponse

ArchivedTriggerActionRequest

An archived TriggerActionRequest

ArchivedTriggerActionResponse

An archived TriggerActionResponse

ArchivedTypingRequest

An archived TypingRequest

ArchivedTypingResponse

An archived TypingResponse

ArchivedUnbanUserRequest

An archived UnbanUserRequest

ArchivedUnbanUserResponse

An archived UnbanUserResponse

ArchivedUnpinMessageRequest

An archived UnpinMessageRequest

ArchivedUnpinMessageResponse

An archived UnpinMessageResponse

ArchivedUpdateAllChannelOrderRequest

An archived UpdateAllChannelOrderRequest

ArchivedUpdateAllChannelOrderResponse

An archived UpdateAllChannelOrderResponse

ArchivedUpdateChannelInformationRequest

An archived UpdateChannelInformationRequest

ArchivedUpdateChannelInformationResponse

An archived UpdateChannelInformationResponse

ArchivedUpdateChannelOrderRequest

An archived UpdateChannelOrderRequest

ArchivedUpdateChannelOrderResponse

An archived UpdateChannelOrderResponse

ArchivedUpdateGuildInformationRequest

An archived UpdateGuildInformationRequest

ArchivedUpdateGuildInformationResponse

An archived UpdateGuildInformationResponse

ArchivedUpdateMessageTextRequest

An archived UpdateMessageTextRequest

ArchivedUpdateMessageTextResponse

An archived UpdateMessageTextResponse

ArchivedUpgradeRoomToGuildRequest

An archived UpgradeRoomToGuildRequest

ArchivedUpgradeRoomToGuildResponse

An archived UpgradeRoomToGuildResponse

Attachment

Object representing a generic message attachment.

AttachmentResolver

The resolver for an archived Attachment

BanUserRequest

Used in BanUser endpoint.

BanUserRequestResolver

The resolver for an archived BanUserRequest

BanUserResponse

Used in BanUser endpoint.

BanUserResponseResolver

The resolver for an archived BanUserResponse

Channel

An object representing a channel, without the ID.

ChannelResolver

The resolver for an archived Channel

ChannelWithId

The channel alongside with an ID.

ChannelWithIdResolver

The resolver for an archived ChannelWithId

Content

Object representing a message’s content.

ContentResolver

The resolver for an archived Content

CreateChannelRequest

Channel Kinds:

CreateChannelRequestResolver

The resolver for an archived CreateChannelRequest

CreateChannelResponse

Used in the CreateChannel endpoint.

CreateChannelResponseResolver

The resolver for an archived CreateChannelResponse

CreateDirectMessageRequest

Used in the CreateDirectMessage endpoint.

CreateDirectMessageRequestResolver

The resolver for an archived CreateDirectMessageRequest

CreateDirectMessageResponse

Used in the CreateDirectMessage endpoint.

CreateDirectMessageResponseResolver

The resolver for an archived CreateDirectMessageResponse

CreateGuildRequest

Request type used in CreateGuild endpoint.

CreateGuildRequestResolver

The resolver for an archived CreateGuildRequest

CreateGuildResponse

Used in the CreateGuild endpoint.

CreateGuildResponseResolver

The resolver for an archived CreateGuildResponse

CreateInviteRequest

Used in the CreateInvite endpoint.

CreateInviteRequestResolver

The resolver for an archived CreateInviteRequest

CreateInviteResponse

Used in the CreateInvite endpoint.

CreateInviteResponseResolver

The resolver for an archived CreateInviteResponse

CreateRoomRequest

Request type used in CreateRoom endpoint.

CreateRoomRequestResolver

The resolver for an archived CreateRoomRequest

CreateRoomResponse

Used in the CreateRoom endpoint.

CreateRoomResponseResolver

The resolver for an archived CreateRoomResponse

DeleteChannelRequest

Used in the DeleteChannel endpoint.

DeleteChannelRequestResolver

The resolver for an archived DeleteChannelRequest

DeleteChannelResponse

Used in the DeleteChannel endpoint.

DeleteChannelResponseResolver

The resolver for an archived DeleteChannelResponse

DeleteGuildRequest

Used in the DeleteGuild endpoint.

DeleteGuildRequestResolver

The resolver for an archived DeleteGuildRequest

DeleteGuildResponse

Used in the DeleteGuild endpoint.

DeleteGuildResponseResolver

The resolver for an archived DeleteGuildResponse

DeleteGuildRoleRequest

Used in the DeleteGuildRole endpoint.

DeleteGuildRoleRequestResolver

The resolver for an archived DeleteGuildRoleRequest

DeleteGuildRoleResponse

Used in the DeleteGuildRole endpoint.

DeleteGuildRoleResponseResolver

The resolver for an archived DeleteGuildRoleResponse

DeleteInviteRequest

Used in the DeleteInvite endpoint.

DeleteInviteRequestResolver

The resolver for an archived DeleteInviteRequest

DeleteInviteResponse

Used in the DeleteInvite endpoint.

DeleteInviteResponseResolver

The resolver for an archived DeleteInviteResponse

DeleteMessageRequest

Used in the DeleteMessage endpoint.

DeleteMessageRequestResolver

The resolver for an archived DeleteMessageRequest

DeleteMessageResponse

Used in the DeleteMessage endpoint.

DeleteMessageResponseResolver

The resolver for an archived DeleteMessageResponse

Embed

Object representing a message embed.

EmbedResolver

The resolver for an archived Embed

Format

A format for text

FormatResolver

The resolver for an archived Format

FormattedText

Formatted text

FormattedTextResolver

The resolver for an archived FormattedText

GetBannedUsersRequest

Used in GetBannedUsers endpoint.

GetBannedUsersRequestResolver

The resolver for an archived GetBannedUsersRequest

GetBannedUsersResponse

Used in GetBannedUsers endpoint.

GetBannedUsersResponseResolver

The resolver for an archived GetBannedUsersResponse

GetChannelMessagesRequest

Used in the GetChannelMessages endpoint.

GetChannelMessagesRequestResolver

The resolver for an archived GetChannelMessagesRequest

GetChannelMessagesResponse

Used in the GetChannelMessages endpoint.

GetChannelMessagesResponseResolver

The resolver for an archived GetChannelMessagesResponse

GetGuildChannelsRequest

Used in the GetGuildChannels endpoint.

GetGuildChannelsRequestResolver

The resolver for an archived GetGuildChannelsRequest

GetGuildChannelsResponse

Used in the GetGuildChannels endpoint.

GetGuildChannelsResponseResolver

The resolver for an archived GetGuildChannelsResponse

GetGuildInvitesRequest

Used in the GetGuildInvites endpoint.

GetGuildInvitesRequestResolver

The resolver for an archived GetGuildInvitesRequest

GetGuildInvitesResponse

Used in the GetGuildInvites endpoint.

GetGuildInvitesResponseResolver

The resolver for an archived GetGuildInvitesResponse

GetGuildListRequest

Used in the GetGuildList endpoint.

GetGuildListRequestResolver

The resolver for an archived GetGuildListRequest

GetGuildListResponse

Used in the GetGuildList endpoint.

GetGuildListResponseResolver

The resolver for an archived GetGuildListResponse

GetGuildMembersRequest

Used in the GetGuildMembers endpoint.

GetGuildMembersRequestResolver

The resolver for an archived GetGuildMembersRequest

GetGuildMembersResponse

Used in the GetGuildMembers endpoint.

GetGuildMembersResponseResolver

The resolver for an archived GetGuildMembersResponse

GetGuildRequest

Used in the GetGuild endpoint.

GetGuildRequestResolver

The resolver for an archived GetGuildRequest

GetGuildResponse

Used in the GetGuild endpoint.

GetGuildResponseResolver

The resolver for an archived GetGuildResponse

GetGuildRolesRequest

Used in the GetGuildRoles endpoint.

GetGuildRolesRequestResolver

The resolver for an archived GetGuildRolesRequest

GetGuildRolesResponse

Used in the GetGuildRoles endpoint.

GetGuildRolesResponseResolver

The resolver for an archived GetGuildRolesResponse

GetMessageRequest

Used in the GetMessage endpoint.

GetMessageRequestResolver

The resolver for an archived GetMessageRequest

GetMessageResponse

Used in the GetMessage endpoint.

GetMessageResponseResolver

The resolver for an archived GetMessageResponse

GetPendingInvitesRequest

Used in GetPendingInvites endpoint.

GetPendingInvitesRequestResolver

The resolver for an archived GetPendingInvitesRequest

GetPendingInvitesResponse

Used in GetPendingInvites endpoint.

GetPendingInvitesResponseResolver

The resolver for an archived GetPendingInvitesResponse

GetPermissionsRequest

Used in the GetPermissions endpoint.

GetPermissionsRequestResolver

The resolver for an archived GetPermissionsRequest

GetPermissionsResponse

Used in the GetPermissions endpoint.

GetPermissionsResponseResolver

The resolver for an archived GetPermissionsResponse

GetPinnedMessagesRequest

Used in the GetPinnedMessages endpoint.

GetPinnedMessagesRequestResolver

The resolver for an archived GetPinnedMessagesRequest

GetPinnedMessagesResponse

Used in the GetPinnedMessages endpoint.

GetPinnedMessagesResponseResolver

The resolver for an archived GetPinnedMessagesResponse

GetUserRolesRequest

Used in the GetUserRoles endpoint.

GetUserRolesRequestResolver

The resolver for an archived GetUserRolesRequest

GetUserRolesResponse

Used in the GetUserRoles endpoint.

GetUserRolesResponseResolver

The resolver for an archived GetUserRolesResponse

GiveUpOwnershipRequest

Request for GiveUpOwnership

GiveUpOwnershipRequestResolver

The resolver for an archived GiveUpOwnershipRequest

GiveUpOwnershipResponse

Response for GiveUpOwnership

GiveUpOwnershipResponseResolver

The resolver for an archived GiveUpOwnershipResponse

GrantOwnershipRequest

Request for GrantOwnership

GrantOwnershipRequestResolver

The resolver for an archived GrantOwnershipRequest

GrantOwnershipResponse

Response for GrantOwnership

GrantOwnershipResponseResolver

The resolver for an archived GrantOwnershipResponse

Guild

Object representing a guild without the ID part.

GuildKind

The kind of a guild.

GuildKindResolver

The resolver for an archived GuildKind

GuildListEntry

Object representing a guild list entry.

GuildListEntryResolver

The resolver for an archived GuildListEntry

GuildResolver

The resolver for an archived Guild

GuildWithId

Object representing a guild with the ID part.

GuildWithIdResolver

The resolver for an archived GuildWithId

IgnorePendingInviteRequest

Used in IgnorePendingInvite endpoint.

IgnorePendingInviteRequestResolver

The resolver for an archived IgnorePendingInviteRequest

IgnorePendingInviteResponse

Used in IgnorePendingInvite endpoint.

IgnorePendingInviteResponseResolver

The resolver for an archived IgnorePendingInviteResponse

Invite

Object representing an invite without the ID part.

InviteResolver

The resolver for an archived Invite

InviteUserToGuildRequest

Used in InviteUserToGuild endpoint.

InviteUserToGuildRequestResolver

The resolver for an archived InviteUserToGuildRequest

InviteUserToGuildResponse

Used in InviteUserToGuild endpoint.

InviteUserToGuildResponseResolver

The resolver for an archived InviteUserToGuildResponse

InviteWithId

Invite with ID.

InviteWithIdResolver

The resolver for an archived InviteWithId

JoinGuildRequest

Used in the JoinGuild endpoint.

JoinGuildRequestResolver

The resolver for an archived JoinGuildRequest

JoinGuildResponse

Used in the JoinGuild endpoint.

JoinGuildResponseResolver

The resolver for an archived JoinGuildResponse

KickUserRequest

Used in KickUser endpoint.

KickUserRequestResolver

The resolver for an archived KickUserRequest

KickUserResponse

Used in KickUser endpoint.

KickUserResponseResolver

The resolver for an archived KickUserResponse

LeaveGuildRequest

Used in the LeaveGuild endpoint.

LeaveGuildRequestResolver

The resolver for an archived LeaveGuildRequest

LeaveGuildResponse

Used in the LeaveGuild endpoint.

LeaveGuildResponseResolver

The resolver for an archived LeaveGuildResponse

ManageUserRolesRequest

Used in the ManageUserRoles endpoint.

ManageUserRolesRequestResolver

The resolver for an archived ManageUserRolesRequest

ManageUserRolesResponse

Used in the ManageUserRoles endpoint.

ManageUserRolesResponseResolver

The resolver for an archived ManageUserRolesResponse

Message

Object representing a message without the ID part.

MessageResolver

The resolver for an archived Message

MessageWithId

Object representing a message with it’s ID.

MessageWithIdResolver

The resolver for an archived MessageWithId

Minithumbnail

Minithumbnail is an extremely low-quality JPEG thumbnail.

MinithumbnailResolver

The resolver for an archived Minithumbnail

ModifyGuildRoleRequest

Used in the ModifyGuildRole endpoint.

ModifyGuildRoleRequestResolver

The resolver for an archived ModifyGuildRoleRequest

ModifyGuildRoleResponse

Used in the ModifyGuildRole endpoint.

ModifyGuildRoleResponseResolver

The resolver for an archived ModifyGuildRoleResponse

MoveRoleRequest

Used in the MoveRole endpoint.

MoveRoleRequestResolver

The resolver for an archived MoveRoleRequest

MoveRoleResponse

Used in the MoveRole endpoint.

MoveRoleResponseResolver

The resolver for an archived MoveRoleResponse

Overrides

Overrides provide a way to override the name and avatar of a message.

OverridesResolver

The resolver for an archived Overrides

PendingInvite

A pending invite.

PendingInviteResolver

The resolver for an archived PendingInvite

Permission

Object representing a single permission node.

PermissionResolver

The resolver for an archived Permission

Photo

Photo contains data about a photo.

PhotoResolver

The resolver for an archived Photo

PinMessageRequest

Used in the PinMessage endpoint.

PinMessageRequestResolver

The resolver for an archived PinMessageRequest

PinMessageResponse

Used in the UnpinMessage endpoint.

PinMessageResponseResolver

The resolver for an archived PinMessageResponse

PreviewGuildRequest

Used in the PreviewGuild endpoint.

PreviewGuildRequestResolver

The resolver for an archived PreviewGuildRequest

PreviewGuildResponse

Used in the PreviewGuild endpoint.

PreviewGuildResponseResolver

The resolver for an archived PreviewGuildResponse

QueryHasPermissionRequest

Used in the QueryHasPermission endpoint.

QueryHasPermissionRequestResolver

The resolver for an archived QueryHasPermissionRequest

QueryHasPermissionResponse

Used in the QueryHasPermission endpoint.

QueryHasPermissionResponseResolver

The resolver for an archived QueryHasPermissionResponse

Reaction

Object representing a reaction.

ReactionResolver

The resolver for an archived Reaction

RejectPendingInviteRequest

Used in RejectPendingInvite endpoint.

RejectPendingInviteRequestResolver

The resolver for an archived RejectPendingInviteRequest

RejectPendingInviteResponse

Used in RejectPendingInvite endpoint.

RejectPendingInviteResponseResolver

The resolver for an archived RejectPendingInviteResponse

RemoveReactionRequest

Used in RemoveReaction endpoint.

RemoveReactionRequestResolver

The resolver for an archived RemoveReactionRequest

RemoveReactionResponse

Used in RemoveReaction endpoint.

RemoveReactionResponseResolver

The resolver for an archived RemoveReactionResponse

Role

Object representing a role without the ID.

RoleResolver

The resolver for an archived Role

RoleWithId

Object representing a role with it’s ID.

RoleWithIdResolver

The resolver for an archived RoleWithId

SendMessageRequest

Used in the SendMessage endpoint.

SendMessageRequestResolver

The resolver for an archived SendMessageRequest

SendMessageResponse

Used in the SendMessage endpoint.

SendMessageResponseResolver

The resolver for an archived SendMessageResponse

SetPermissionsRequest

Used in the SetPermissions endpoint.

SetPermissionsRequestResolver

The resolver for an archived SetPermissionsRequest

SetPermissionsResponse

Used in the SetPermissions endpoint.

SetPermissionsResponseResolver

The resolver for an archived SetPermissionsResponse

StreamEvent

Describes an event.

StreamEventResolver

The resolver for an archived StreamEvent

StreamEventsRequest

Request type for use in the StreamEvents endpoint. By default, this endpoint will subscribe to all events. Any guilds joined in the future will be added to the subscription as well. Use the UnsubscribeFromAll event for unsubscribing from all current subscriptions and disable the automatic guild subscriptions

StreamEventsRequestResolver

The resolver for an archived StreamEventsRequest

StreamEventsResponse

Used in the StreamEvents endpoint.

StreamEventsResponseResolver

The resolver for an archived StreamEventsResponse

TriggerActionRequest

Used in the TriggerAction endpoint.

TriggerActionRequestResolver

The resolver for an archived TriggerActionRequest

TriggerActionResponse

Used in the TriggerAction endpoint.

TriggerActionResponseResolver

The resolver for an archived TriggerActionResponse

TypingRequest

Used in Typing endpoint.

TypingRequestResolver

The resolver for an archived TypingRequest

TypingResponse

Used in Typing endpoint.

TypingResponseResolver

The resolver for an archived TypingResponse

UnbanUserRequest

Used in UnbanUser endpoint.

UnbanUserRequestResolver

The resolver for an archived UnbanUserRequest

UnbanUserResponse

Used in UnbanUser endpoint.

UnbanUserResponseResolver

The resolver for an archived UnbanUserResponse

UnpinMessageRequest

Used in the UnpinMessage endpoint.

UnpinMessageRequestResolver

The resolver for an archived UnpinMessageRequest

UnpinMessageResponse

Used in the UnpinMessage endpoint.

UnpinMessageResponseResolver

The resolver for an archived UnpinMessageResponse

UpdateAllChannelOrderRequest

Request specifiying the order of all channels in a guild at once

UpdateAllChannelOrderRequestResolver

The resolver for an archived UpdateAllChannelOrderRequest

UpdateAllChannelOrderResponse

Used in the UpdateAllChannelOrder endpoint.

UpdateAllChannelOrderResponseResolver

The resolver for an archived UpdateAllChannelOrderResponse

UpdateChannelInformationRequest

Used in the UpdateChannelInformation endpoint.

UpdateChannelInformationRequestResolver

The resolver for an archived UpdateChannelInformationRequest

UpdateChannelInformationResponse

Used in the UpdateChannelInformation endpoint.

UpdateChannelInformationResponseResolver

The resolver for an archived UpdateChannelInformationResponse

UpdateChannelOrderRequest

Used in the UpdateChannelOrder endpoint.

UpdateChannelOrderRequestResolver

The resolver for an archived UpdateChannelOrderRequest

UpdateChannelOrderResponse

Used in the UpdateChannelOrder endpoint.

UpdateChannelOrderResponseResolver

The resolver for an archived UpdateChannelOrderResponse

UpdateGuildInformationRequest

Used in the UpdateGuildInformation endpoint.

UpdateGuildInformationRequestResolver

The resolver for an archived UpdateGuildInformationRequest

UpdateGuildInformationResponse

Used in the UpdateGuildInformation endpoint.

UpdateGuildInformationResponseResolver

The resolver for an archived UpdateGuildInformationResponse

UpdateMessageTextRequest

Used in the UpdateMessageText endpoint.

UpdateMessageTextRequestResolver

The resolver for an archived UpdateMessageTextRequest

UpdateMessageTextResponse

Used in the UpdateMessageText endpoint.

UpdateMessageTextResponseResolver

The resolver for an archived UpdateMessageTextResponse

UpgradeRoomToGuildRequest

Used in the UpgradeRoomToGuild endpoint.

UpgradeRoomToGuildRequestResolver

The resolver for an archived UpgradeRoomToGuildRequest

UpgradeRoomToGuildResponse

Used in the UpgradeRoomToGuild endpoint.

UpgradeRoomToGuildResponseResolver

The resolver for an archived UpgradeRoomToGuildResponse

Enums

ArchivedChannelKind

An archived ChannelKind

ArchivedLeaveReason

An archived LeaveReason

ChannelKind

What kind the channel is.

ChannelKindResolver

The resolver for an archived ChannelKind

LeaveReason

A reason for why a user has left a guild.

LeaveReasonResolver

The resolver for an archived LeaveReason