Module harmony_rust_sdk::api::chat::v1::action

Nested message and enum types in Action.

Modules

dropdown

Nested message and enum types in Dropdown.

Structs

ArchivedButton

An archived Button

ArchivedDropdown

An archived Dropdown

ArchivedInput

An archived Input

Button

A button that users can click on to trigger an action.

ButtonResolver

The resolver for an archived Button

Dropdown

A dropdown menu that users can click on to trigger an action.

DropdownResolver

The resolver for an archived Dropdown

Input

A text input that users can type in to trigger an action.

InputResolver

The resolver for an archived Input

Enums

ArchivedKind

An archived Kind

ArchivedType

An archived Type

Kind

The kind of the action.

KindResolver

The resolver for an archived Kind

Type

The action type. This is primarily used to change the look of the action to the user (example: Destructive actions will have a red background).

TypeResolver

The resolver for an archived Type