Struct harmony_rust_sdk::api::chat::v1::chat_service_client::ChatServiceClient

pub struct ChatServiceClient { /* fields omitted */ }

The core of Harmony’s chat operations.

Implementations

impl ChatServiceClient

pub fn new_inner(inner: Client) -> Self

pub fn new(inner: ReqwestClient, host_url: Url) -> ClientResult<Self>

pub fn inner(&self) -> &Client

pub async fn create_guild(
&mut self,
request: impl IntoRequest<CreateGuildRequest>
) -> ClientResult<CreateGuildResponse>

Endpoint to create a guild.

pub async fn create_invite(
&mut self,
request: impl IntoRequest<CreateInviteRequest>
) -> ClientResult<CreateInviteResponse>

Endpoint to create an invite.

pub async fn create_channel(
&mut self,
request: impl IntoRequest<CreateChannelRequest>
) -> ClientResult<CreateChannelResponse>

Endpoint to create a channel.

pub async fn get_guild_list(
&mut self,
request: impl IntoRequest<GetGuildListRequest>
) -> ClientResult<GetGuildListResponse>

Endpoint to get your guild list.

pub async fn get_guild(
&mut self,
request: impl IntoRequest<GetGuildRequest>
) -> ClientResult<GetGuildResponse>

Endpoint to get information about a guild.

pub async fn get_guild_invites(
&mut self,
request: impl IntoRequest<GetGuildInvitesRequest>
) -> ClientResult<GetGuildInvitesResponse>

Endpoint to get the invites of a guild.

This requires the “invites.view” permission.

pub async fn get_guild_members(
&mut self,
request: impl IntoRequest<GetGuildMembersRequest>
) -> ClientResult<GetGuildMembersResponse>

Endpoint to get the members of a guild.

pub async fn get_guild_channels(
&mut self,
request: impl IntoRequest<GetGuildChannelsRequest>
) -> ClientResult<GetGuildChannelsResponse>

Endpoint to get the channels of a guild.

You will only be informed of channels you have the “messages.view” permission for.

pub async fn get_channel_messages(
&mut self,
request: impl IntoRequest<GetChannelMessagesRequest>
) -> ClientResult<GetChannelMessagesResponse>

Endpoint to get the messages from a guild channel.

pub async fn get_message(
&mut self,
request: impl IntoRequest<GetMessageRequest>
) -> ClientResult<GetMessageResponse>

Endpoint to get a single message from a guild channel.

pub async fn update_guild_information(
&mut self,
request: impl IntoRequest<UpdateGuildInformationRequest>
) -> ClientResult<UpdateGuildInformationResponse>

Endpoint to update a guild’s information.

pub async fn update_channel_information(
&mut self,
request: impl IntoRequest<UpdateChannelInformationRequest>
) -> ClientResult<UpdateChannelInformationResponse>

Endpoint to update a channel’s information.

pub async fn update_channel_order(
&mut self,
request: impl IntoRequest<UpdateChannelOrderRequest>
) -> ClientResult<UpdateChannelOrderResponse>

Endpoint to change the position of a channel in the channel list.

pub async fn update_all_channel_order(
&mut self,
request: impl IntoRequest<UpdateAllChannelOrderRequest>
) -> ClientResult<UpdateAllChannelOrderResponse>

Endpoint to change the position of all channels in the guild; must specifcy all channels or fails

pub async fn update_message_text(
&mut self,
request: impl IntoRequest<UpdateMessageTextRequest>
) -> ClientResult<UpdateMessageTextResponse>

Endpoint to change the text of a message.

pub async fn delete_guild(
&mut self,
request: impl IntoRequest<DeleteGuildRequest>
) -> ClientResult<DeleteGuildResponse>

Endpoint to delete a guild.

pub async fn delete_invite(
&mut self,
request: impl IntoRequest<DeleteInviteRequest>
) -> ClientResult<DeleteInviteResponse>

Endpoint to delete an invite.

pub async fn delete_channel(
&mut self,
request: impl IntoRequest<DeleteChannelRequest>
) -> ClientResult<DeleteChannelResponse>

Endpoint to delete a channel.

pub async fn delete_message(
&mut self,
request: impl IntoRequest<DeleteMessageRequest>
) -> ClientResult<DeleteMessageResponse>

Endpoint to delete a message.

This requires the “messages.manage.delete” permission if you are not the message author.

pub async fn join_guild(
&mut self,
request: impl IntoRequest<JoinGuildRequest>
) -> ClientResult<JoinGuildResponse>

Endpoint to join a guild.

pub async fn leave_guild(
&mut self,
request: impl IntoRequest<LeaveGuildRequest>
) -> ClientResult<LeaveGuildResponse>

Endpoint to leave a guild.

pub async fn trigger_action(
&mut self,
request: impl IntoRequest<TriggerActionRequest>
) -> ClientResult<TriggerActionResponse>

Endpont to trigger an action.

pub async fn send_message(
&mut self,
request: impl IntoRequest<SendMessageRequest>
) -> ClientResult<SendMessageResponse>

Endpoint to send a message.

pub async fn query_has_permission(
&mut self,
request: impl IntoRequest<QueryHasPermissionRequest>
) -> ClientResult<QueryHasPermissionResponse>

Endpoint to query if a user has a permission.

pub async fn set_permissions(
&mut self,
request: impl IntoRequest<SetPermissionsRequest>
) -> ClientResult<SetPermissionsResponse>

Endpoint to set permissions for a role in a guild.

pub async fn get_permissions(
&mut self,
request: impl IntoRequest<GetPermissionsRequest>
) -> ClientResult<GetPermissionsResponse>

Endpoint to get permissions for a role in a guild.

pub async fn move_role(
&mut self,
request: impl IntoRequest<MoveRoleRequest>
) -> ClientResult<MoveRoleResponse>

Endpoint to change the position of a role in the role list in a guild.

pub async fn get_guild_roles(
&mut self,
request: impl IntoRequest<GetGuildRolesRequest>
) -> ClientResult<GetGuildRolesResponse>

Endpoint to get the roles from a guild.

pub async fn add_guild_role(
&mut self,
request: impl IntoRequest<AddGuildRoleRequest>
) -> ClientResult<AddGuildRoleResponse>

Endpoint to add a role to a guild.

pub async fn modify_guild_role(
&mut self,
request: impl IntoRequest<ModifyGuildRoleRequest>
) -> ClientResult<ModifyGuildRoleResponse>

Endpoint to a modify a role from a guild.

pub async fn delete_guild_role(
&mut self,
request: impl IntoRequest<DeleteGuildRoleRequest>
) -> ClientResult<DeleteGuildRoleResponse>

Endpoint to delete a role from a guild.

pub async fn manage_user_roles(
&mut self,
request: impl IntoRequest<ManageUserRolesRequest>
) -> ClientResult<ManageUserRolesResponse>

Endpoint to manage the roles of a guild member.

pub async fn get_user_roles(
&mut self,
request: impl IntoRequest<GetUserRolesRequest>
) -> ClientResult<GetUserRolesResponse>

Endpoint to get the roles a guild member has.

pub async fn typing(
&mut self,
request: impl IntoRequest<TypingRequest>
) -> ClientResult<TypingResponse>

Endpoint to send a typing notification in a guild channel.

pub async fn preview_guild(
&mut self,
request: impl IntoRequest<PreviewGuildRequest>
) -> ClientResult<PreviewGuildResponse>

Endpoint to “preview” a guild, which returns some information about a guild.

pub async fn get_banned_users(
&mut self,
request: impl IntoRequest<GetBannedUsersRequest>
) -> ClientResult<GetBannedUsersResponse>

Endpoint to get banned users in a guild.

pub async fn ban_user(
&mut self,
request: impl IntoRequest<BanUserRequest>
) -> ClientResult<BanUserResponse>

Endpoint to ban a user from a guild.

pub async fn kick_user(
&mut self,
request: impl IntoRequest<KickUserRequest>
) -> ClientResult<KickUserResponse>

Endpoint to kick a user from a guild.

pub async fn unban_user(
&mut self,
request: impl IntoRequest<UnbanUserRequest>
) -> ClientResult<UnbanUserResponse>

Endpoint to unban a user from a guild.

pub async fn get_pinned_messages(
&mut self,
request: impl IntoRequest<GetPinnedMessagesRequest>
) -> ClientResult<GetPinnedMessagesResponse>

Endpoint to get all pinned messages in a guild channel.

pub async fn pin_message(
&mut self,
request: impl IntoRequest<PinMessageRequest>
) -> ClientResult<PinMessageResponse>

Endpoint to pin a message in a guild channel.

pub async fn unpin_message(
&mut self,
request: impl IntoRequest<UnpinMessageRequest>
) -> ClientResult<UnpinMessageResponse>

Endpoint to unpin a message in a guild channel.

pub async fn stream_events(
&mut self,
request: impl IntoRequest<()>
) -> ClientResult<Socket<StreamEventsRequest, StreamEventsResponse>>

Endpoint to stream events from the homeserver.

pub async fn add_reaction(
&mut self,
request: impl IntoRequest<AddReactionRequest>
) -> ClientResult<AddReactionResponse>

Endpoint to add a reaction to a message.

pub async fn remove_reaction(
&mut self,
request: impl IntoRequest<RemoveReactionRequest>
) -> ClientResult<RemoveReactionResponse>

Endpoint to remove a reaction from a message.