Module harmony_rust_sdk::api::chat::v1::event

Nested message and enum types in Event.

Structs

ActionPerformed
Event sent when an action is performed.

ChannelCreated
Event sent when a new channel is created.

ChannelDeleted
Event sent when a channel is deleted.

ChannelUpdated
Event sent when a channel’s information or position in the channel is changed.

ChannelsReordered
Event sent when all channels have been reordered

EmotePackAdded
Event sent when an emote pack is added.

EmotePackDeleted
Event sent when an emote pack is deleted.

EmotePackEmotesUpdated
Event sent when an emote pack’s emotes were changed.

EmotePackUpdated
Event sent when an emote pack’s information is changed.

GuildAddedToList
Event sent when you join a new guild.

GuildDeleted
Event sent when a guild is deleted.

GuildRemovedFromList
Event sent when you leave a guild.

GuildUpdated
Event sent when a guild’s information is updated.

MemberJoined
Event sent a user joins to a guild.

MemberLeft
Event sent when a member of a guild leaves said guild for whatever reason.

MessageDeleted
Event sent when a message is deleted.

MessageSent
Event sent when a new message is sent.

MessageUpdated
Event sent when a message’s text content is updated.

PermissionUpdated
Event sent when a permission is changed that matters to you.

ProfileUpdated
Event sent when a user’s profile is updated.

RoleCreated
Event sent when a role is created.

RoleDeleted
Event sent when a role is deleted.

RoleMoved
Event sent when a role’s position in the role list is changed.

RolePermissionsUpdated
Event sent when a role’s permissions are changed.

RoleUpdated
Event sent when a role’s information is changed.

Typing
Event sent when a user sends a typing notification in a guild channel.

UserRolesUpdated
Event sent when a user’s roles are changed.

Enums

Event
Which event to send.

LeaveReason
A reason for why a user has left a guild.