Module harmony_rust_sdk::api::chat::v1::format

Nested message and enum types in Format.

Modules

color

Nested message and enum types in Color.

Structs

Bold

Bold text

ChannelMention

Mention of a channel (on the current guild)

CodeBlock

A larger codeblock, with a programming language specified Clients should ideally not bound the width of codeblock messages, possibly scrolling the codeblock horizontally separately from the rest of the message

Color

Colour modification

Emoji

An emoji

GuildMention

Mention of a guild

Italic

Italic text

Localization

Replace a part of the text with the text matching the i18n code. If i18n code was not found, keep the original text.

Monospace

Monospace text

RoleMention

Mention of a role (on the current guild)

Subscript

Subscript text

Superscript

Superscript text

Underline

Underlined text

UserMention

Mention of a user (on the current homeserver)

Enums

Format

the style if format to apply to this text