Module harmony_rust_sdk::api::chat::v1::stream_event

Nested message and enum types in StreamEvent.

Structs

ActionPerformed

Event sent when an action is performed.

ActionPerformedResolver

The resolver for an archived ActionPerformed

ArchivedActionPerformed

An archived ActionPerformed

ArchivedChannelCreated

An archived ChannelCreated

ArchivedChannelDeleted

An archived ChannelDeleted

ArchivedChannelPositionUpdated

An archived ChannelPositionUpdated

ArchivedChannelUpdated

An archived ChannelUpdated

ArchivedChannelsReordered

An archived ChannelsReordered

ArchivedGuildAddedToList

An archived GuildAddedToList

ArchivedGuildDeleted

An archived GuildDeleted

ArchivedGuildRemovedFromList

An archived GuildRemovedFromList

ArchivedGuildUpdated

An archived GuildUpdated

ArchivedInviteReceived

An archived InviteReceived

ArchivedInviteRejected

An archived InviteRejected

ArchivedMemberJoined

An archived MemberJoined

ArchivedMemberLeft

An archived MemberLeft

ArchivedMessageDeleted

An archived MessageDeleted

ArchivedMessagePinned

An archived MessagePinned

ArchivedMessageSent

An archived MessageSent

ArchivedMessageUnpinned

An archived MessageUnpinned

ArchivedMessageUpdated

An archived MessageUpdated

ArchivedOwnerAdded

An archived OwnerAdded

ArchivedOwnerRemoved

An archived OwnerRemoved

ArchivedPermissionUpdated

An archived PermissionUpdated

ArchivedReactionUpdated

An archived ReactionUpdated

ArchivedRoleCreated

An archived RoleCreated

ArchivedRoleDeleted

An archived RoleDeleted

ArchivedRoleMoved

An archived RoleMoved

ArchivedRolePermissionsUpdated

An archived RolePermissionsUpdated

ArchivedRoleUpdated

An archived RoleUpdated

ArchivedTyping

An archived Typing

ArchivedUserRolesUpdated

An archived UserRolesUpdated

ChannelCreated

Event sent when a new channel is created.

ChannelCreatedResolver

The resolver for an archived ChannelCreated

ChannelDeleted

Event sent when a channel is deleted.

ChannelDeletedResolver

The resolver for an archived ChannelDeleted

ChannelPositionUpdated

Event sent when a channel’s position in the channel list is changed.

ChannelPositionUpdatedResolver

The resolver for an archived ChannelPositionUpdated

ChannelUpdated

Event sent when a channel’s information is changed.

ChannelUpdatedResolver

The resolver for an archived ChannelUpdated

ChannelsReordered

Event sent when all channels have been reordered

ChannelsReorderedResolver

The resolver for an archived ChannelsReordered

GuildAddedToList

Event sent when you join a new guild.

GuildAddedToListResolver

The resolver for an archived GuildAddedToList

GuildDeleted

Event sent when a guild is deleted.

GuildDeletedResolver

The resolver for an archived GuildDeleted

GuildRemovedFromList

Event sent when you leave a guild.

GuildRemovedFromListResolver

The resolver for an archived GuildRemovedFromList

GuildUpdated

Event sent when a guild’s information is changed.

GuildUpdatedResolver

The resolver for an archived GuildUpdated

InviteReceived

Sent when a guild invite is received.

InviteReceivedResolver

The resolver for an archived InviteReceived

InviteRejected

Sent when a guild invite is rejected by the invitee.

InviteRejectedResolver

The resolver for an archived InviteRejected

MemberJoined

Event sent a user joins to a guild.

MemberJoinedResolver

The resolver for an archived MemberJoined

MemberLeft

Event sent when a member of a guild leaves said guild for whatever reason.

MemberLeftResolver

The resolver for an archived MemberLeft

MessageDeleted

Event sent when a message is deleted.

MessageDeletedResolver

The resolver for an archived MessageDeleted

MessagePinned

Sent when a message is pinned in a guild channel.

MessagePinnedResolver

The resolver for an archived MessagePinned

MessageSent

Event sent when a new message is sent.

MessageSentResolver

The resolver for an archived MessageSent

MessageUnpinned

Sent when a message is unpinned in a guild channel.

MessageUnpinnedResolver

The resolver for an archived MessageUnpinned

MessageUpdated

Event sent when a message’s text content is updated.

MessageUpdatedResolver

The resolver for an archived MessageUpdated

OwnerAdded

Sent when there’s a new owner.

OwnerAddedResolver

The resolver for an archived OwnerAdded

OwnerRemoved

Sent when an owner gives up their ownership.

OwnerRemovedResolver

The resolver for an archived OwnerRemoved

PermissionUpdated

Event sent when a permission is changed that matters to you.

PermissionUpdatedResolver

The resolver for an archived PermissionUpdated

ReactionUpdated

Sent when a message’s reaction is changed.

ReactionUpdatedResolver

The resolver for an archived ReactionUpdated

RoleCreated

Event sent when a role is created.

RoleCreatedResolver

The resolver for an archived RoleCreated

RoleDeleted

Event sent when a role is deleted.

RoleDeletedResolver

The resolver for an archived RoleDeleted

RoleMoved

Event sent when a role’s position in the role list is changed.

RoleMovedResolver

The resolver for an archived RoleMoved

RolePermissionsUpdated

Event sent when a role’s permissions are changed.

RolePermissionsUpdatedResolver

The resolver for an archived RolePermissionsUpdated

RoleUpdated

Event sent when a role’s information is changed.

RoleUpdatedResolver

The resolver for an archived RoleUpdated

Typing

Event sent when a user sends a typing notification in a guild channel.

TypingResolver

The resolver for an archived Typing

UserRolesUpdated

Event sent when a user’s roles are changed.

UserRolesUpdatedResolver

The resolver for an archived UserRolesUpdated

Enums

ArchivedEvent

An archived Event

Event

Which event to send.

EventResolver

The resolver for an archived Event