Trait harmony_rust_sdk::api::emote::v1::emote_service_server::EmoteService

pub trait EmoteService: Send + Sync + 'static {
    type Error: CustomError + Send + Sync + 'static;

    #[must_use]
    fn create_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<CreateEmotePackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_emote_packs<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetEmotePacksRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetEmotePacksResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_emote_pack_emotes<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetEmotePackEmotesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetEmotePackEmotesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn add_emote_to_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<AddEmoteToPackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<AddEmoteToPackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteEmotePackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_emote_from_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteEmoteFromPackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteEmoteFromPackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn dequip_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DequipEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DequipEmotePackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn equip_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<EquipEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<EquipEmotePackResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn create_emote_pack_pre(&self) -> BoxedFilter<()> { ... }

    fn get_emote_packs_pre(&self) -> BoxedFilter<()> { ... }

    fn get_emote_pack_emotes_pre(&self) -> BoxedFilter<()> { ... }

    fn add_emote_to_pack_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_emote_pack_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_emote_from_pack_pre(&self) -> BoxedFilter<()> { ... }

    fn dequip_emote_pack_pre(&self) -> BoxedFilter<()> { ... }

    fn equip_emote_pack_pre(&self) -> BoxedFilter<()> { ... }

}

Generated trait containing hRPC methods that should be implemented for use with EmoteServiceServer.

Associated Types

type Error: CustomError + Send + Sync + 'static

Loading content...
Required methods

#[must_use]
fn create_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<CreateEmotePackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to create an emote pack.

#[must_use]
fn get_emote_packs<'life0, 'async_trait>(
&'life0 self,
request: Request<GetEmotePacksRequest>
) -> Pin<Box<dyn Future<Output = Result<GetEmotePacksResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the emote packs you have equipped.

#[must_use]
fn get_emote_pack_emotes<'life0, 'async_trait>(
&'life0 self,
request: Request<GetEmotePackEmotesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetEmotePackEmotesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the emotes in an emote pack.

#[must_use]
fn add_emote_to_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<AddEmoteToPackRequest>
) -> Pin<Box<dyn Future<Output = Result<AddEmoteToPackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to add an emote to an emote pack that you own.

#[must_use]
fn delete_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteEmotePackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete an emote pack that you own.

#[must_use]
fn delete_emote_from_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteEmoteFromPackRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteEmoteFromPackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete an emote from an emote pack.

#[must_use]
fn dequip_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DequipEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<DequipEmotePackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to dequip an emote pack that you have equipped.

#[must_use]
fn equip_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<EquipEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<EquipEmotePackResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to equip an emote pack.

Loading content...
Provided methods

fn create_emote_pack_pre(&self) -> BoxedFilter<()>

fn get_emote_packs_pre(&self) -> BoxedFilter<()>

fn get_emote_pack_emotes_pre(&self) -> BoxedFilter<()>

fn add_emote_to_pack_pre(&self) -> BoxedFilter<()>

fn delete_emote_pack_pre(&self) -> BoxedFilter<()>

fn delete_emote_from_pack_pre(&self) -> BoxedFilter<()>

fn dequip_emote_pack_pre(&self) -> BoxedFilter<()>

fn equip_emote_pack_pre(&self) -> BoxedFilter<()>

Loading content...