Module harmony_rust_sdk::api::harmonytypes::v1

Modules

override

Structs

Action
Rules on actions: Actions without a type are assumed to be Button actions or Menu actions when nested. Button actions can only have Menu actions as children. Dropdown actions can only have non-recursive Menu actions as children. Menu actions can only have other Menu actions as children. SmallEntry and LargeActions cannot have children.

Attachment
TYPES

Embed

EmbedField

EmbedHeading

HarmonyMethodMetadata

Message

Metadata

Override
OVERRIDES

Enums

ActionPresentation

ActionType

FieldPresentation

UserStatus