Trait harmony_rust_sdk::api::mediaproxy::v1::media_proxy_service_server::MediaProxyService

pub trait MediaProxyService: Sync + Send + 'static {
    fn fetch_link_metadata(
        &self, 
        request: HrpcRequest<FetchLinkMetadataRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<FetchLinkMetadataResponse>>>;

    fn instant_view(
        &self, 
        request: HrpcRequest<InstantViewRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<InstantViewResponse>>>;

    fn can_instant_view(
        &self, 
        request: HrpcRequest<CanInstantViewRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CanInstantViewResponse>>>;

    fn fetch_link_metadata_middleware(&self) -> Option<HrpcLayer> { ... }

    fn instant_view_middleware(&self) -> Option<HrpcLayer> { ... }

    fn can_instant_view_middleware(&self) -> Option<HrpcLayer> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with MediaProxyServiceServer.

Required methods

fn fetch_link_metadata(
&self,
request: HrpcRequest<FetchLinkMetadataRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<FetchLinkMetadataResponse>>>

Endpoint to fetch metadata for a URL.

fn instant_view(
&self,
request: HrpcRequest<InstantViewRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<InstantViewResponse>>>

Endpoint to instant view a website URL.

fn can_instant_view(
&self,
request: HrpcRequest<CanInstantViewRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CanInstantViewResponse>>>

Endpoint to query if the server can generate an instant view for a website URL.

Provided methods

fn fetch_link_metadata_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn instant_view_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn can_instant_view_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.