Module harmony_rust_sdk::api::profile::v1

v1 of profile service.

Modules

profile_override

Nested message and enum types in ProfileOverride.

profile_service_client

Generated client implementations.

stream_event

Nested message and enum types in StreamEvent.

Structs

AppDataOverrides

The message used for the ‘h.overrides’ key of appdata.

GetAppDataRequest

Used in GetAppData endpoint.

GetAppDataResponse

Used in GetAppData endpoint.

GetProfileRequest

Used in GetProfile endpoint.

GetProfileResponse

Used in GetProfile endpoint.

OverrideTag

A tag for an override. This is used as a standard shorthand for sending a message with an override. If a message starts with before and ends with after, clients should send a message with the override the tag belongs to, stripping the tag indicators.

Profile

Data for a single profile, without the user’s ID.

ProfileOverride

An individual override

ProfileUpdated

Event sent when a user’s profile is updated.

SetAppDataRequest

Used in SetAppData endpoint.

SetAppDataResponse

Used in SetAppData endpoint.

StreamEvent

Describes an emote service event.

UpdateProfileRequest

Used in UpdateProfile endpoint.

UpdateProfileResponse

Used in UpdateProfile endpoint.

Enums

AccountKind

The possible kinds of an account

UserStatus

The possible statuses a user can have.