Trait harmony_rust_sdk::api::profile::v1::profile_service_server::ProfileService

pub trait ProfileService: Sync + Send + 'static {
    fn get_profile(
        &self, 
        request: HrpcRequest<GetProfileRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetProfileResponse>>>;

    fn update_profile(
        &self, 
        request: HrpcRequest<UpdateProfileRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateProfileResponse>>>;

    fn get_app_data(
        &self, 
        request: HrpcRequest<GetAppDataRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetAppDataResponse>>>;

    fn set_app_data(
        &self, 
        request: HrpcRequest<SetAppDataRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<SetAppDataResponse>>>;

    fn get_profile_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_profile_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_app_data_middleware(&self) -> Option<HrpcLayer> { ... }

    fn set_app_data_middleware(&self) -> Option<HrpcLayer> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with ProfileServiceServer.

Required methods

fn get_profile(
&self,
request: HrpcRequest<GetProfileRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetProfileResponse>>>

Gets a user’s profile.

fn update_profile(
&self,
request: HrpcRequest<UpdateProfileRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateProfileResponse>>>

Updates the user’s profile.

fn get_app_data(
&self,
request: HrpcRequest<GetAppDataRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetAppDataResponse>>>

Gets app data for a user (this can be used to store user preferences which is synchronized across devices).

fn set_app_data(
&self,
request: HrpcRequest<SetAppDataRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<SetAppDataResponse>>>

Sets the app data for a user.

Provided methods

fn get_profile_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_profile_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_app_data_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn set_app_data_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.