Trait harmony_rust_sdk::api::sync::v1::postbox_service_server::PostboxService

pub trait PostboxService: Sync + Send + 'static {
    fn pull(
        &self, 
        request: HrpcRequest<PullRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<PullResponse>>>;

    fn push(
        &self, 
        request: HrpcRequest<PushRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<PushResponse>>>;

    fn notify_new_id(
        &self, 
        request: HrpcRequest<NotifyNewIdRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<NotifyNewIdResponse>>>;

    fn pull_middleware(&self) -> Option<HrpcLayer> { ... }

    fn push_middleware(&self) -> Option<HrpcLayer> { ... }

    fn notify_new_id_middleware(&self) -> Option<HrpcLayer> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with PostboxServiceServer.

Required methods

fn pull(
&self,
request: HrpcRequest<PullRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<PullResponse>>>

Endpoint to pull events.

fn push(
&self,
request: HrpcRequest<PushRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<PushResponse>>>

Endpoint to push events.

fn notify_new_id(
&self,
request: HrpcRequest<NotifyNewIdRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<NotifyNewIdResponse>>>

Endpoint to notify a server of a server ID change. It is called by the server that had it’s server ID changed for all servers it has federated with.

Provided methods

fn pull_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn push_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn notify_new_id_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.