Module harmony_rust_sdk::api::voice::v1::stream_message_response

Nested message and enum types in StreamMessageResponse.

Structs

ArchivedInitialized

An archived Initialized

ArchivedJoinedChannel

An archived JoinedChannel

ArchivedPreparedForJoinChannel

An archived PreparedForJoinChannel

ArchivedUserJoined

An archived UserJoined

ArchivedUserLeft

An archived UserLeft

Initialized

Initialization data for client

InitializedResolver

The resolver for an archived Initialized

JoinedChannel

Producer for voice created; consumer and producer transports are connected

JoinedChannelResolver

The resolver for an archived JoinedChannel

PreparedForJoinChannel

RTP capabilities validated

PreparedForJoinChannelResolver

The resolver for an archived PreparedForJoinChannel

UserJoined

Data for the user that joined the room and it’s producer

UserJoinedResolver

The resolver for an archived UserJoined

UserLeft

Data for the user that left the room and the producer

UserLeftResolver

The resolver for an archived UserLeft

Enums

ArchivedMessage

An archived Message

Message

Message for this response

MessageResolver

The resolver for an archived Message