Module harmony_rust_sdk::client

Rust client implementation for Harmony, powered by [hrpc].

See the examples directory in the repository on how to use this.

Modules

api
Client API implementations.

error
Error related code used by Client.

exports
Some crates exported for user convenience.

Structs

AuthSocket
Auth steps subscription socket.

Client
Client implementation for Harmony.

EventsSocket
Event subscription socket.

Enums

AuthStatus
Represents an authentication state in which a Client can be.