Module harmony_rust_sdk::client::api::chat

Chat service client API.

Re-exports

pub use crate::api::chat::event;

pub use crate::api::chat::get_emote_pack_emotes_response::Emote;

pub use crate::api::chat::get_emote_packs_response::EmotePack;

pub use crate::api::chat::get_guild_channels_response::Channel;

pub use crate::api::chat::get_guild_invites_response::Invite;

pub use crate::api::chat::get_guild_list_response::GuildListEntry;

pub use crate::api::chat::permission::Mode;

pub use crate::api::chat::stream_events_request;

pub use crate::api::chat::Event;

pub use crate::api::chat::InviteId;

pub use crate::api::chat::PermissionList;

pub use crate::api::chat::Place;

pub use crate::api::chat::Role;

Modules

channel
Manage and query channels.

emote
Manage and query emotes and emote packs.

guild
Manage and query guilds.

invite
Manage and query invites.

message
Manage and query messages.

permissions
Manage and query user permissions and roles.

profile
Manage and query user profiles.

Enums

EventSource
Describes where to subscribe for events.

Functions

stream_events
Streams events from specified event sources from the server.

trigger_action
Triggers the specified action.

typing
Notifies the server that the local user is typing.