Module harmony_rust_sdk::client::api::chat::invite

Manage and query invites.

Re-exports

pub use crate::api::chat::get_guild_invites_response;

pub use crate::api::chat::CreateInviteRequest;

pub use crate::api::chat::DeleteInviteRequest;

pub use crate::api::chat::GetGuildInvitesRequest;

pub use crate::api::chat::InviteId;

Structs

CreateInvite
Convenience type to create a valid CreateInviteRequest.

DeleteInvite
Convenience type to create a valid DeleteInviteRequest.

Functions

create_invite
Create an invite with a name and a number of uses for the specified guild.

delete_invite
Delete an invite with the specified name in the specified guild.

get_guild_invites
Get invites for the specified guild.