Trait harmony_rust_sdk::client::api::chat::message::MessageExt

pub trait MessageExt {
    fn text(&self) -> Option<&str>;

    fn embeds(&self) -> Option<&Embed>;

    fn files(&self) -> Option<&[Attachment]>;
}

Trait that implements convenience methods for Message type.

Required methods

fn text(&self) -> Option<&str>

Get the text content of the message if it has one.

fn embeds(&self) -> Option<&Embed>

Get the embed content of the message if it has one.

fn files(&self) -> Option<&[Attachment]>

Get the file content of the message if it has one.

Loading content...