Module harmony_rust_sdk::client::api::rest

REST client API.

Structs

DownloadedFile
A downloaded file.

Enums

FileId
A "file id", which can either be a HMC URL or a plain ID string.

FileKind
Kind of the file downloaded.

Functions

download
Downloads a file using a file ID.

download_extract_file
Downloads a file then extracts file information from it.

upload
Uploads a file to the homeserver.

upload_extract_id
Uploads a file and then extracts the file ID from the returned response.