Module harmony_rust_sdk::client::error

Error related code used by Client.

Re-exports

pub use crate::api::HmcParseError;

Structs

ReqwestError

The Errors that may occur when processing a Request.

UrlError

An error resulting from a failed attempt to construct a URI.

Enums

ClientError

Error type used by Client.

SocketError

Errors that can occur during a socket’s operation.

Type Definitions

ClientResult

Result type used by many Client methods.

InternalClientError