Struct harmony_rust_sdk::client::Client

pub struct Client { /* fields omitted */ }

Client implementation for Harmony.

Implementations

impl Client

pub async fn new(
homeserver_url: Url,
session: Option<Session>
) -> ClientResult<Self>

Create a new Client from a homeserver Url (URL) and an (optional) session.

If port is not specified in the URL, this will add the default port 2289 to it. If scheme is not specified, this will assume the scheme is https.

Example

let client = Client::new("https://chat.harmonyapp.io:2289".parse().unwrap(), None).await?;

pub async fn event_loop<'a, Fut, Hndlr>(
&'a self,
subs: Vec<EventSource>,
handler: Hndlr
) -> Result<(), ClientError> where
Fut: Future<Output = ClientResult<bool>> + 'a,
Hndlr: Fn(&'a Client, Event) -> Fut,

Consumes self and starts an event loop with the given handler.

All socket errors will be logged with tracing. If the handler function returns Ok(true) or Err(err), the function will return, so if you don’t want it to return, return Ok(false).

pub async fn chat(&self) -> MutexGuard<'_, ChatServiceClient>

Get a mutex guard to the chat service.

pub async fn auth(&self) -> MutexGuard<'_, AuthServiceClient>

Get a mutex guard to the auth service.

pub async fn mediaproxy(&self) -> MutexGuard<'_, MediaProxyServiceClient>

Get a mutex guard to the mediaproxy service.

pub async fn profile(&self) -> MutexGuard<'_, ProfileServiceClient>

Get a mutex guard to the profile service.

pub async fn emote(&self) -> MutexGuard<'_, EmoteServiceClient>

Get a mutex guard to the emote service.

pub async fn call<Req: CallRequest>(
&self,
request: Req
) -> ClientResult<Req::Response>

Execute the given request.

pub fn auth_status(&self) -> AuthStatus

Get the current auth status.

Example

assert!(!client.auth_status().is_authenticated());

pub fn homeserver_url(&self) -> &Url

Get the stored homeserver URL.

Example

assert_eq!(&client.homeserver_url().to_string(), "https://chat.harmonyapp.io:2289/");

pub fn make_hmc(&self, id: impl Display) -> Result<Hmc, HmcFromStrError>

Makes an HMC with homeserver’s authority and the given ID.

Example

assert_eq!(client.make_hmc("404").unwrap(), Hmc::new("chat.harmonyapp.io:2289", "404").unwrap());

pub async fn begin_auth(&self) -> ClientResult<()>

Start an authentication session.

Example

client.begin_auth().await?;
// Do auth stuff here

pub async fn next_auth_step(
&self,
response: AuthStepResponse
) -> ClientResult<Option<NextStepResponse>>

Request the next authentication step from the server.

Returns Ok(None) if authentication was completed. Returns Ok(Some(AuthStep)) if extra step is requested from the server.

Example

client.begin_auth().await?;
let next_step = client.next_auth_step(AuthStepResponse::Initial).await?;
// Do more auth stuff here

pub async fn prev_auth_step(&self) -> ClientResult<StepBackResponse>

Go back to the previous authentication step.

Example

client.begin_auth().await?;
// Call next step and whatnot here
// Oops, user wants to do something else, lets go back
let prev_step = client.prev_auth_step().await?;
// Do more auth stuff here

pub async fn auth_stream(&self) -> ClientResult<AuthSocket>

Begin an authentication steps stream for the current authentication session.

Example

client.begin_auth().await?;
let auth_steps_stream = client.auth_stream().await?;
// Do auth stuff here

pub async fn subscribe_events(
&self,
subscriptions: Vec<EventSource>
) -> ClientResult<EventsSocket>

Subscribe to events coming from specified event sources.

Example

// Auth here
let event_stream = client.subscribe_events(vec![EventSource::Homeserver, EventSource::Action]).await?;
// Do more auth stuff here