Module harmony_rust_sdk::api::auth::v1

Modules

auth_service_client
Generated client implementations.

auth_service_server
Generated server implementations.

auth_step
Nested message and enum types in AuthStep.

next_step_request
Nested message and enum types in NextStepRequest.

Structs

AuthStep
AuthStep A step in the authentication process Contains a variety of different types of views It is recommended to have a fallback_url specified For non-trivial authentication procedures (such as captchas)

BeginAuthResponse

FederateReply

FederateRequest
Connection

KeyReply

LoginFederatedRequest

NextStepRequest
NextStepRequest contains the client’s response to the server’s challenge This needs to be called first with no arguments to receive the first step

Session

StepBackRequest
StepBackRequest A request to go back 1 step

StreamStepsRequest
StreamStepsRequest Required to be initiated by all authenticating clients Allows the server to send steps