Module harmony_rust_sdk::api::auth::v1::auth_step

Nested message and enum types in AuthStep.

Modules

form
Nested message and enum types in Form.

Structs

Choice
Choice A step which allows the user to choose from a range of options Allows you to show a heading by specifying title

Form
Form A step which requires the user to input information Allows you to show a heading by specifying title

Waiting
Waiting A step which requires the user to perform an external action The title and description should explain to the user what they should do to complete this step

Enums

Step