Module harmony_rust_sdk::api::chat::v1

Modules

action

Nested message and enum types in Action.

action_payload

Nested message and enum types in ActionPayload.

all_permissions

All chat permissions.

chat_service_client

Generated client implementations.

content

Nested message and enum types in Content.

embed

Nested message and enum types in Embed.

format

Nested message and enum types in Format.

get_channel_messages_request

Nested message and enum types in GetChannelMessagesRequest.

guild_kind

Nested message and enum types in GuildKind.

overrides

Nested message and enum types in Overrides.

stream_event

Nested message and enum types in StreamEvent.

stream_events_request

Nested message and enum types in StreamEventsRequest.

stream_events_response

Nested message and enum types in StreamEventsResponse.

Structs

Action

Actions are interactive elements that can exist within an embed.

ActionPayload

The payload sent to the bot when an action is triggered.

AddGuildRoleRequest

Used in the AddGuildRole endpoint.

AddGuildRoleResponse

Used in the AddGuildRole endpoint.

AddReactionRequest

Used in AddReaction endpoint.

AddReactionResponse

Used in AddReaction endpoint.

Attachment

Object representing a generic message attachment.

BanUserRequest

Used in BanUser endpoint.

BanUserResponse

Used in BanUser endpoint.

Channel

An object representing a channel, without the ID.

ChannelWithId

The channel alongside with an ID.

Content

Object representing a message’s content.

CreateChannelRequest

Channel Kinds:

CreateChannelResponse

Used in the CreateChannel endpoint.

CreateDirectMessageRequest

Used in the CreateDirectMessage endpoint.

CreateDirectMessageResponse

Used in the CreateDirectMessage endpoint.

CreateGuildRequest

Request type used in CreateGuild endpoint.

CreateGuildResponse

Used in the CreateGuild endpoint.

CreateInviteRequest

Used in the CreateInvite endpoint.

CreateInviteResponse

Used in the CreateInvite endpoint.

CreateRoomRequest

Request type used in CreateRoom endpoint.

CreateRoomResponse

Used in the CreateRoom endpoint.

DeleteChannelRequest

Used in the DeleteChannel endpoint.

DeleteChannelResponse

Used in the DeleteChannel endpoint.

DeleteGuildRequest

Used in the DeleteGuild endpoint.

DeleteGuildResponse

Used in the DeleteGuild endpoint.

DeleteGuildRoleRequest

Used in the DeleteGuildRole endpoint.

DeleteGuildRoleResponse

Used in the DeleteGuildRole endpoint.

DeleteInviteRequest

Used in the DeleteInvite endpoint.

DeleteInviteResponse

Used in the DeleteInvite endpoint.

DeleteMessageRequest

Used in the DeleteMessage endpoint.

DeleteMessageResponse

Used in the DeleteMessage endpoint.

Embed

Object representing a message embed.

Format

A format for text

FormattedText

Formatted text

GetBannedUsersRequest

Used in GetBannedUsers endpoint.

GetBannedUsersResponse

Used in GetBannedUsers endpoint.

GetChannelMessagesRequest

Used in the GetChannelMessages endpoint.

GetChannelMessagesResponse

Used in the GetChannelMessages endpoint.

GetGuildChannelsRequest

Used in the GetGuildChannels endpoint.

GetGuildChannelsResponse

Used in the GetGuildChannels endpoint.

GetGuildInvitesRequest

Used in the GetGuildInvites endpoint.

GetGuildInvitesResponse

Used in the GetGuildInvites endpoint.

GetGuildListRequest

Used in the GetGuildList endpoint.

GetGuildListResponse

Used in the GetGuildList endpoint.

GetGuildMembersRequest

Used in the GetGuildMembers endpoint.

GetGuildMembersResponse

Used in the GetGuildMembers endpoint.

GetGuildRequest

Used in the GetGuild endpoint.

GetGuildResponse

Used in the GetGuild endpoint.

GetGuildRolesRequest

Used in the GetGuildRoles endpoint.

GetGuildRolesResponse

Used in the GetGuildRoles endpoint.

GetMessageRequest

Used in the GetMessage endpoint.

GetMessageResponse

Used in the GetMessage endpoint.

GetPendingInvitesRequest

Used in GetPendingInvites endpoint.

GetPendingInvitesResponse

Used in GetPendingInvites endpoint.

GetPermissionsRequest

Used in the GetPermissions endpoint.

GetPermissionsResponse

Used in the GetPermissions endpoint.

GetPinnedMessagesRequest

Used in the GetPinnedMessages endpoint.

GetPinnedMessagesResponse

Used in the GetPinnedMessages endpoint.

GetUserRolesRequest

Used in the GetUserRoles endpoint.

GetUserRolesResponse

Used in the GetUserRoles endpoint.

GiveUpOwnershipRequest

Request for GiveUpOwnership

GiveUpOwnershipResponse

Response for GiveUpOwnership

GrantOwnershipRequest

Request for GrantOwnership

GrantOwnershipResponse

Response for GrantOwnership

Guild

Object representing a guild without the ID part.

GuildKind

The kind of a guild.

GuildListEntry

Object representing a guild list entry.

GuildWithId

Object representing a guild with the ID part.

IgnorePendingInviteRequest

Used in IgnorePendingInvite endpoint.

IgnorePendingInviteResponse

Used in IgnorePendingInvite endpoint.

Invite

Object representing an invite without the ID part.

InviteUserToGuildRequest

Used in InviteUserToGuild endpoint.

InviteUserToGuildResponse

Used in InviteUserToGuild endpoint.

InviteWithId

Invite with ID.

JoinGuildRequest

Used in the JoinGuild endpoint.

JoinGuildResponse

Used in the JoinGuild endpoint.

KickUserRequest

Used in KickUser endpoint.

KickUserResponse

Used in KickUser endpoint.

LeaveGuildRequest

Used in the LeaveGuild endpoint.

LeaveGuildResponse

Used in the LeaveGuild endpoint.

ManageUserRolesRequest

Used in the ManageUserRoles endpoint.

ManageUserRolesResponse

Used in the ManageUserRoles endpoint.

Message

Object representing a message without the ID part.

MessageWithId

Object representing a message with it’s ID.

Minithumbnail

Minithumbnail is an extremely low-quality JPEG thumbnail.

ModifyGuildRoleRequest

Used in the ModifyGuildRole endpoint.

ModifyGuildRoleResponse

Used in the ModifyGuildRole endpoint.

MoveRoleRequest

Used in the MoveRole endpoint.

MoveRoleResponse

Used in the MoveRole endpoint.

Overrides

Overrides provide a way to override the name and avatar of a message.

PendingInvite

A pending invite.

Permission

Object representing a single permission node.

Photo

Photo contains data about a photo.

PinMessageRequest

Used in the PinMessage endpoint.

PinMessageResponse

Used in the UnpinMessage endpoint.

PreviewGuildRequest

Used in the PreviewGuild endpoint.

PreviewGuildResponse

Used in the PreviewGuild endpoint.

QueryHasPermissionRequest

Used in the QueryHasPermission endpoint.

QueryHasPermissionResponse

Used in the QueryHasPermission endpoint.

Reaction

Object representing a reaction.

RejectPendingInviteRequest

Used in RejectPendingInvite endpoint.

RejectPendingInviteResponse

Used in RejectPendingInvite endpoint.

RemoveReactionRequest

Used in RemoveReaction endpoint.

RemoveReactionResponse

Used in RemoveReaction endpoint.

Role

Object representing a role without the ID.

RoleWithId

Object representing a role with it’s ID.

SendMessageRequest

Used in the SendMessage endpoint.

SendMessageResponse

Used in the SendMessage endpoint.

SetPermissionsRequest

Used in the SetPermissions endpoint.

SetPermissionsResponse

Used in the SetPermissions endpoint.

StreamEvent

Describes an event.

StreamEventsRequest

Request type for use in the StreamEvents endpoint. By default, this endpoint will subscribe to all events. Any guilds joined in the future will be added to the subscription as well. Use the UnsubscribeFromAll event for unsubscribing from all current subscriptions and disable the automatic guild subscriptions

StreamEventsResponse

Used in the StreamEvents endpoint.

TriggerActionRequest

Used in the TriggerAction endpoint.

TriggerActionResponse

Used in the TriggerAction endpoint.

TypingRequest

Used in Typing endpoint.

TypingResponse

Used in Typing endpoint.

UnbanUserRequest

Used in UnbanUser endpoint.

UnbanUserResponse

Used in UnbanUser endpoint.

UnpinMessageRequest

Used in the UnpinMessage endpoint.

UnpinMessageResponse

Used in the UnpinMessage endpoint.

UpdateAllChannelOrderRequest

Request specifiying the order of all channels in a guild at once

UpdateAllChannelOrderResponse

Used in the UpdateAllChannelOrder endpoint.

UpdateChannelInformationRequest

Used in the UpdateChannelInformation endpoint.

UpdateChannelInformationResponse

Used in the UpdateChannelInformation endpoint.

UpdateChannelOrderRequest

Used in the UpdateChannelOrder endpoint.

UpdateChannelOrderResponse

Used in the UpdateChannelOrder endpoint.

UpdateGuildInformationRequest

Used in the UpdateGuildInformation endpoint.

UpdateGuildInformationResponse

Used in the UpdateGuildInformation endpoint.

UpdateMessageTextRequest

Used in the UpdateMessageText endpoint.

UpdateMessageTextResponse

Used in the UpdateMessageText endpoint.

UpgradeRoomToGuildRequest

Used in the UpgradeRoomToGuild endpoint.

UpgradeRoomToGuildResponse

Used in the UpgradeRoomToGuild endpoint.

Enums

ChannelKind

What kind the channel is.

LeaveReason

A reason for why a user has left a guild.