Struct harmony_rust_sdk::api::chat::v1::chat_service_client::ChatServiceClient

pub struct ChatServiceClient<Inner> { /* fields omitted */ }

The core of Harmony’s chat operations.

Implementations

impl<Inner, InnerErr> ChatServiceClient<Inner> where
Inner: Service<TransportRequest, Response = TransportResponse, Error = TransportError<InnerErr>> + 'static,
InnerErr: 'static,

pub fn create_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<CreateGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<CreateGuildRequest>,

Endpoint to create a guild.

pub fn create_room<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<CreateRoomResponse>, InnerErr>> + 'static where
Req: IntoRequest<CreateRoomRequest>,

Endpoint to create a “room” guild.

pub fn create_direct_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<CreateDirectMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<CreateDirectMessageRequest>,

Endpoint to create a “direct message” guild.

The inviter and the invitee that join the created guild will both be owners.

The guild should be created on the server that inviter is on.

The server should send a guild invite to the invitee (specified in the request), using the UserInvited postbox event if the invitee is on another server, otherwise see the below item.

The server should process this as follows: adding the invite to their pending invite list and sending an InviteReceived event over their event stream if the invitee is on this server.

The invitee may or may not use the invite. Only the invitee may use the invite.

pub fn upgrade_room_to_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpgradeRoomToGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpgradeRoomToGuildRequest>,

Endpoint to upgrade a “room” guild to a “normal” guild.

pub fn create_invite<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<CreateInviteResponse>, InnerErr>> + 'static where
Req: IntoRequest<CreateInviteRequest>,

Endpoint to create an invite.

pub fn create_channel<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<CreateChannelResponse>, InnerErr>> + 'static where
Req: IntoRequest<CreateChannelRequest>,

Endpoint to create a channel.

pub fn get_guild_list<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildListResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildListRequest>,

Endpoint to get your guild list.

pub fn invite_user_to_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<InviteUserToGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<InviteUserToGuildRequest>,

Endpoint to invite a user to a guild.

pub fn get_pending_invites<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetPendingInvitesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetPendingInvitesRequest>,

Endpoint to get your pending invites.

pub fn reject_pending_invite<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<RejectPendingInviteResponse>, InnerErr>> + 'static where
Req: IntoRequest<RejectPendingInviteRequest>,

Endpoint to reject (delete with notification to the inviter) an invite from your pending invite list.

If the invitee is on a different server than the inviter, the invitee’s server should send a UserRejectedInvite postbox event to the inviter’s server.

The “notification” is sending a InviteRejected stream event to the inviter. If the guild’s kind is DirectMessage, the server should also set the rejected field of the guild’s kind to true.

pub fn ignore_pending_invite<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<IgnorePendingInviteResponse>, InnerErr>> + 'static where
Req: IntoRequest<IgnorePendingInviteRequest>,

Endpoint to ignore (delete without notification to the inviter) an invite from your pending invite list.

pub fn get_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildRequest>,

Endpoint to get information about a guild.

pub fn get_guild_invites<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildInvitesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildInvitesRequest>,

Endpoint to get the invites of a guild.

This requires the “invites.view” permission.

pub fn get_guild_members<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildMembersResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildMembersRequest>,

Endpoint to get the members of a guild.

pub fn get_guild_channels<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildChannelsResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildChannelsRequest>,

Endpoint to get the channels of a guild.

You will only be informed of channels you have the “messages.view” permission for.

pub fn get_channel_messages<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetChannelMessagesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetChannelMessagesRequest>,

Endpoint to get the messages from a guild channel.

pub fn get_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetMessageRequest>,

Endpoint to get a single message from a guild channel.

pub fn update_guild_information<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpdateGuildInformationResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpdateGuildInformationRequest>,

Endpoint to update a guild’s information.

pub fn update_channel_information<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpdateChannelInformationResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpdateChannelInformationRequest>,

Endpoint to update a channel’s information.

pub fn update_channel_order<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpdateChannelOrderResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpdateChannelOrderRequest>,

Endpoint to change the position of a channel in the channel list.

pub fn update_all_channel_order<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpdateAllChannelOrderResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpdateAllChannelOrderRequest>,

Endpoint to change the position of all channels in the guild; must specifcy all channels or fails

pub fn update_message_text<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UpdateMessageTextResponse>, InnerErr>> + 'static where
Req: IntoRequest<UpdateMessageTextRequest>,

Endpoint to change the text of a message.

pub fn delete_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<DeleteGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<DeleteGuildRequest>,

Endpoint to delete a guild. Can only be invoked if there’s one owner.

pub fn delete_invite<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<DeleteInviteResponse>, InnerErr>> + 'static where
Req: IntoRequest<DeleteInviteRequest>,

Endpoint to delete an invite.

pub fn delete_channel<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<DeleteChannelResponse>, InnerErr>> + 'static where
Req: IntoRequest<DeleteChannelRequest>,

Endpoint to delete a channel.

pub fn delete_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<DeleteMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<DeleteMessageRequest>,

Endpoint to delete a message.

This requires the “messages.manage.delete” permission if you are not the message author.

pub fn join_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<JoinGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<JoinGuildRequest>,

Endpoint to join a guild.

If the invite used is in a user’s “pending invites” list, it should be removed from there.

pub fn leave_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<LeaveGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<LeaveGuildRequest>,

Endpoint to leave a guild.

If you’re the only owner, you can’t leave a guild. Exception to this rule are “direct message” guilds.

When the last member in a “direct message” guild leaves it, that guild should be deleted.

pub fn trigger_action<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<TriggerActionResponse>, InnerErr>> + 'static where
Req: IntoRequest<TriggerActionRequest>,

Endpont to trigger an action.

pub fn send_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<SendMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<SendMessageRequest>,

Endpoint to send a message.

pub fn query_has_permission<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<QueryHasPermissionResponse>, InnerErr>> + 'static where
Req: IntoRequest<QueryHasPermissionRequest>,

Endpoint to query if a user has a permission.

pub fn set_permissions<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<SetPermissionsResponse>, InnerErr>> + 'static where
Req: IntoRequest<SetPermissionsRequest>,

Endpoint to set permissions for a role in a guild.

pub fn get_permissions<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetPermissionsResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetPermissionsRequest>,

Endpoint to get permissions for a role in a guild.

pub fn move_role<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<MoveRoleResponse>, InnerErr>> + 'static where
Req: IntoRequest<MoveRoleRequest>,

Endpoint to change the position of a role in the role list in a guild.

pub fn get_guild_roles<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetGuildRolesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetGuildRolesRequest>,

Endpoint to get the roles from a guild.

pub fn add_guild_role<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<AddGuildRoleResponse>, InnerErr>> + 'static where
Req: IntoRequest<AddGuildRoleRequest>,

Endpoint to add a role to a guild.

pub fn modify_guild_role<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<ModifyGuildRoleResponse>, InnerErr>> + 'static where
Req: IntoRequest<ModifyGuildRoleRequest>,

Endpoint to a modify a role from a guild.

pub fn delete_guild_role<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<DeleteGuildRoleResponse>, InnerErr>> + 'static where
Req: IntoRequest<DeleteGuildRoleRequest>,

Endpoint to delete a role from a guild.

pub fn manage_user_roles<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<ManageUserRolesResponse>, InnerErr>> + 'static where
Req: IntoRequest<ManageUserRolesRequest>,

Endpoint to manage the roles of a guild member.

pub fn get_user_roles<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetUserRolesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetUserRolesRequest>,

Endpoint to get the roles a guild member has.

pub fn typing<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<TypingResponse>, InnerErr>> + 'static where
Req: IntoRequest<TypingRequest>,

Endpoint to send a typing notification in a guild channel.

pub fn preview_guild<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<PreviewGuildResponse>, InnerErr>> + 'static where
Req: IntoRequest<PreviewGuildRequest>,

Endpoint to “preview” a guild, which returns some information about a guild.

Guilds with the “direct message” kind can only be previewed by the user who is invited to the guild.

pub fn get_banned_users<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetBannedUsersResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetBannedUsersRequest>,

Endpoint to get banned users in a guild.

pub fn ban_user<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<BanUserResponse>, InnerErr>> + 'static where
Req: IntoRequest<BanUserRequest>,

Endpoint to ban a user from a guild.

pub fn kick_user<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<KickUserResponse>, InnerErr>> + 'static where
Req: IntoRequest<KickUserRequest>,

Endpoint to kick a user from a guild.

pub fn unban_user<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UnbanUserResponse>, InnerErr>> + 'static where
Req: IntoRequest<UnbanUserRequest>,

Endpoint to unban a user from a guild.

pub fn get_pinned_messages<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GetPinnedMessagesResponse>, InnerErr>> + 'static where
Req: IntoRequest<GetPinnedMessagesRequest>,

Endpoint to get all pinned messages in a guild channel.

pub fn pin_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<PinMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<PinMessageRequest>,

Endpoint to pin a message in a guild channel.

pub fn unpin_message<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<UnpinMessageResponse>, InnerErr>> + 'static where
Req: IntoRequest<UnpinMessageRequest>,

Endpoint to unpin a message in a guild channel.

pub fn stream_events<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Socket<StreamEventsRequest, StreamEventsResponse>, InnerErr>> + 'static where
Req: IntoRequest<()>,

Endpoint to stream events from the homeserver.

pub fn add_reaction<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<AddReactionResponse>, InnerErr>> + 'static where
Req: IntoRequest<AddReactionRequest>,

Endpoint to add a reaction to a message.

pub fn remove_reaction<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<RemoveReactionResponse>, InnerErr>> + 'static where
Req: IntoRequest<RemoveReactionRequest>,

Endpoint to remove a reaction from a message.

pub fn grant_ownership<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GrantOwnershipResponse>, InnerErr>> + 'static where
Req: IntoRequest<GrantOwnershipRequest>,

Endpoint to give ownership to someone else.

pub fn give_up_ownership<Req>(
&mut self,
req: Req
) -> impl Future<Output = ClientResult<Response<GiveUpOwnershipResponse>, InnerErr>> + 'static where
Req: IntoRequest<GiveUpOwnershipRequest>,

Endpoint to give up your own ownership. Requires that at least one other person will still be owner.

impl<Inner> ChatServiceClient<Inner>

pub fn new_transport(transport: Inner) -> Self

Create a new client using the provided transport.

pub fn new_inner(client: Client<Inner>) -> Self

Create a new client using the provided generic client.