Struct harmony_rust_sdk::api::chat::v1::chat_service_client::ChatServiceClient

pub struct ChatServiceClient<T> { /* fields omitted */ }

Implementations

impl ChatServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ChatServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_guild<'_>(
&'_ mut self,
request: impl IntoRequest<CreateGuildRequest>
) -> Result<Response<CreateGuildResponse>, Status>

pub async fn create_invite<'_>(
&'_ mut self,
request: impl IntoRequest<CreateInviteRequest>
) -> Result<Response<CreateInviteResponse>, Status>

pub async fn create_channel<'_>(
&'_ mut self,
request: impl IntoRequest<CreateChannelRequest>
) -> Result<Response<CreateChannelResponse>, Status>

pub async fn create_emote_pack<'_>(
&'_ mut self,
request: impl IntoRequest<CreateEmotePackRequest>
) -> Result<Response<CreateEmotePackResponse>, Status>

pub async fn get_guild_list<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildListRequest>
) -> Result<Response<GetGuildListResponse>, Status>

pub async fn add_guild_to_guild_list<'_>(
&'_ mut self,
request: impl IntoRequest<AddGuildToGuildListRequest>
) -> Result<Response<AddGuildToGuildListResponse>, Status>

pub async fn remove_guild_from_guild_list<'_>(
&'_ mut self,
request: impl IntoRequest<RemoveGuildFromGuildListRequest>
) -> Result<Response<RemoveGuildFromGuildListResponse>, Status>

pub async fn get_guild<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildRequest>
) -> Result<Response<GetGuildResponse>, Status>

pub async fn get_guild_invites<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildInvitesRequest>
) -> Result<Response<GetGuildInvitesResponse>, Status>

This requires the "invites.view" permission.

pub async fn get_guild_members<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildMembersRequest>
) -> Result<Response<GetGuildMembersResponse>, Status>

pub async fn get_guild_channels<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildChannelsRequest>
) -> Result<Response<GetGuildChannelsResponse>, Status>

You will only be informed of channels you have the "messages.view" permission for.

pub async fn get_channel_messages<'_>(
&'_ mut self,
request: impl IntoRequest<GetChannelMessagesRequest>
) -> Result<Response<GetChannelMessagesResponse>, Status>

pub async fn get_message<'_>(
&'_ mut self,
request: impl IntoRequest<GetMessageRequest>
) -> Result<Response<GetMessageResponse>, Status>

pub async fn get_emote_packs<'_>(
&'_ mut self,
request: impl IntoRequest<GetEmotePacksRequest>
) -> Result<Response<GetEmotePacksResponse>, Status>

pub async fn get_emote_pack_emotes<'_>(
&'_ mut self,
request: impl IntoRequest<GetEmotePackEmotesRequest>
) -> Result<Response<GetEmotePackEmotesResponse>, Status>

pub async fn update_guild_information<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateGuildInformationRequest>
) -> Result<Response<()>, Status>

pub async fn update_channel_information<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateChannelInformationRequest>
) -> Result<Response<()>, Status>

pub async fn update_channel_order<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateChannelOrderRequest>
) -> Result<Response<()>, Status>

pub async fn update_message<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateMessageRequest>
) -> Result<Response<()>, Status>

pub async fn add_emote_to_pack<'_>(
&'_ mut self,
request: impl IntoRequest<AddEmoteToPackRequest>
) -> Result<Response<()>, Status>

pub async fn delete_guild<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteGuildRequest>
) -> Result<Response<()>, Status>

pub async fn delete_invite<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteInviteRequest>
) -> Result<Response<()>, Status>

pub async fn delete_channel<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteChannelRequest>
) -> Result<Response<()>, Status>

pub async fn delete_message<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteMessageRequest>
) -> Result<Response<()>, Status>

This requires the "messages.manage.delete" permission if you are not the message author.

pub async fn delete_emote_from_pack<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteEmoteFromPackRequest>
) -> Result<Response<()>, Status>

pub async fn delete_emote_pack<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteEmotePackRequest>
) -> Result<Response<()>, Status>

pub async fn dequip_emote_pack<'_>(
&'_ mut self,
request: impl IntoRequest<DequipEmotePackRequest>
) -> Result<Response<()>, Status>

pub async fn join_guild<'_>(
&'_ mut self,
request: impl IntoRequest<JoinGuildRequest>
) -> Result<Response<JoinGuildResponse>, Status>

pub async fn leave_guild<'_>(
&'_ mut self,
request: impl IntoRequest<LeaveGuildRequest>
) -> Result<Response<()>, Status>

pub async fn trigger_action<'_>(
&'_ mut self,
request: impl IntoRequest<TriggerActionRequest>
) -> Result<Response<()>, Status>

pub async fn send_message<'_>(
&'_ mut self,
request: impl IntoRequest<SendMessageRequest>
) -> Result<Response<SendMessageResponse>, Status>

pub async fn query_has_permission<'_>(
&'_ mut self,
request: impl IntoRequest<QueryPermissionsRequest>
) -> Result<Response<QueryPermissionsResponse>, Status>

pub async fn set_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<SetPermissionsRequest>
) -> Result<Response<()>, Status>

pub async fn get_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<GetPermissionsRequest>
) -> Result<Response<GetPermissionsResponse>, Status>

pub async fn move_role<'_>(
&'_ mut self,
request: impl IntoRequest<MoveRoleRequest>
) -> Result<Response<MoveRoleResponse>, Status>

pub async fn get_guild_roles<'_>(
&'_ mut self,
request: impl IntoRequest<GetGuildRolesRequest>
) -> Result<Response<GetGuildRolesResponse>, Status>

pub async fn add_guild_role<'_>(
&'_ mut self,
request: impl IntoRequest<AddGuildRoleRequest>
) -> Result<Response<AddGuildRoleResponse>, Status>

pub async fn modify_guild_role<'_>(
&'_ mut self,
request: impl IntoRequest<ModifyGuildRoleRequest>
) -> Result<Response<()>, Status>

pub async fn delete_guild_role<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteGuildRoleRequest>
) -> Result<Response<()>, Status>

pub async fn manage_user_roles<'_>(
&'_ mut self,
request: impl IntoRequest<ManageUserRolesRequest>
) -> Result<Response<()>, Status>

pub async fn get_user_roles<'_>(
&'_ mut self,
request: impl IntoRequest<GetUserRolesRequest>
) -> Result<Response<GetUserRolesResponse>, Status>

pub async fn stream_events<'_>(
&'_ mut self,
request: impl IntoStreamingRequest<Message = StreamEventsRequest>
) -> Result<Response<Streaming<Event>>, Status>

pub async fn sync<'_>(
&'_ mut self,
request: impl IntoRequest<SyncRequest>
) -> Result<Response<Streaming<SyncEvent>>, Status>

pub async fn get_user<'_>(
&'_ mut self,
request: impl IntoRequest<GetUserRequest>
) -> Result<Response<GetUserResponse>, Status>

pub async fn get_user_metadata<'_>(
&'_ mut self,
request: impl IntoRequest<GetUserMetadataRequest>
) -> Result<Response<GetUserMetadataResponse>, Status>

pub async fn profile_update<'_>(
&'_ mut self,
request: impl IntoRequest<ProfileUpdateRequest>
) -> Result<Response<()>, Status>

pub async fn typing<'_>(
&'_ mut self,
request: impl IntoRequest<TypingRequest>
) -> Result<Response<()>, Status>

pub async fn preview_guild<'_>(
&'_ mut self,
request: impl IntoRequest<PreviewGuildRequest>
) -> Result<Response<PreviewGuildResponse>, Status>