Trait harmony_rust_sdk::api::chat::v1::chat_service_server::ChatService

pub trait ChatService: Sync + Send + 'static {

    fn create_guild(
        &self, 
        request: HrpcRequest<CreateGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateGuildResponse>>>;

    fn create_room(
        &self, 
        request: HrpcRequest<CreateRoomRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateRoomResponse>>>;

    fn create_direct_message(
        &self, 
        request: HrpcRequest<CreateDirectMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateDirectMessageResponse>>>;

    fn upgrade_room_to_guild(
        &self, 
        request: HrpcRequest<UpgradeRoomToGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpgradeRoomToGuildResponse>>>;

    fn create_invite(
        &self, 
        request: HrpcRequest<CreateInviteRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateInviteResponse>>>;

    fn create_channel(
        &self, 
        request: HrpcRequest<CreateChannelRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateChannelResponse>>>;

    fn get_guild_list(
        &self, 
        request: HrpcRequest<GetGuildListRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildListResponse>>>;

    fn invite_user_to_guild(
        &self, 
        request: HrpcRequest<InviteUserToGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<InviteUserToGuildResponse>>>;

    fn get_pending_invites(
        &self, 
        request: HrpcRequest<GetPendingInvitesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPendingInvitesResponse>>>;

    fn reject_pending_invite(
        &self, 
        request: HrpcRequest<RejectPendingInviteRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<RejectPendingInviteResponse>>>;

    fn ignore_pending_invite(
        &self, 
        request: HrpcRequest<IgnorePendingInviteRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<IgnorePendingInviteResponse>>>;

    fn get_guild(
        &self, 
        request: HrpcRequest<GetGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildResponse>>>;

    fn get_guild_invites(
        &self, 
        request: HrpcRequest<GetGuildInvitesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildInvitesResponse>>>;

    fn get_guild_members(
        &self, 
        request: HrpcRequest<GetGuildMembersRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildMembersResponse>>>;

    fn get_guild_channels(
        &self, 
        request: HrpcRequest<GetGuildChannelsRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildChannelsResponse>>>;

    fn get_channel_messages(
        &self, 
        request: HrpcRequest<GetChannelMessagesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetChannelMessagesResponse>>>;

    fn get_message(
        &self, 
        request: HrpcRequest<GetMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetMessageResponse>>>;

    fn update_guild_information(
        &self, 
        request: HrpcRequest<UpdateGuildInformationRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateGuildInformationResponse>>>;

    fn update_channel_information(
        &self, 
        request: HrpcRequest<UpdateChannelInformationRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateChannelInformationResponse>>>;

    fn update_channel_order(
        &self, 
        request: HrpcRequest<UpdateChannelOrderRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateChannelOrderResponse>>>;

    fn update_all_channel_order(
        &self, 
        request: HrpcRequest<UpdateAllChannelOrderRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateAllChannelOrderResponse>>>;

    fn update_message_text(
        &self, 
        request: HrpcRequest<UpdateMessageTextRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateMessageTextResponse>>>;

    fn delete_guild(
        &self, 
        request: HrpcRequest<DeleteGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteGuildResponse>>>;

    fn delete_invite(
        &self, 
        request: HrpcRequest<DeleteInviteRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteInviteResponse>>>;

    fn delete_channel(
        &self, 
        request: HrpcRequest<DeleteChannelRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteChannelResponse>>>;

    fn delete_message(
        &self, 
        request: HrpcRequest<DeleteMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteMessageResponse>>>;

    fn join_guild(
        &self, 
        request: HrpcRequest<JoinGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<JoinGuildResponse>>>;

    fn leave_guild(
        &self, 
        request: HrpcRequest<LeaveGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<LeaveGuildResponse>>>;

    fn trigger_action(
        &self, 
        request: HrpcRequest<TriggerActionRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<TriggerActionResponse>>>;

    fn send_message(
        &self, 
        request: HrpcRequest<SendMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<SendMessageResponse>>>;

    fn query_has_permission(
        &self, 
        request: HrpcRequest<QueryHasPermissionRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<QueryHasPermissionResponse>>>;

    fn set_permissions(
        &self, 
        request: HrpcRequest<SetPermissionsRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<SetPermissionsResponse>>>;

    fn get_permissions(
        &self, 
        request: HrpcRequest<GetPermissionsRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPermissionsResponse>>>;

    fn move_role(
        &self, 
        request: HrpcRequest<MoveRoleRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<MoveRoleResponse>>>;

    fn get_guild_roles(
        &self, 
        request: HrpcRequest<GetGuildRolesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildRolesResponse>>>;

    fn add_guild_role(
        &self, 
        request: HrpcRequest<AddGuildRoleRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<AddGuildRoleResponse>>>;

    fn modify_guild_role(
        &self, 
        request: HrpcRequest<ModifyGuildRoleRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<ModifyGuildRoleResponse>>>;

    fn delete_guild_role(
        &self, 
        request: HrpcRequest<DeleteGuildRoleRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteGuildRoleResponse>>>;

    fn manage_user_roles(
        &self, 
        request: HrpcRequest<ManageUserRolesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<ManageUserRolesResponse>>>;

    fn get_user_roles(
        &self, 
        request: HrpcRequest<GetUserRolesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetUserRolesResponse>>>;

    fn typing(
        &self, 
        request: HrpcRequest<TypingRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<TypingResponse>>>;

    fn preview_guild(
        &self, 
        request: HrpcRequest<PreviewGuildRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<PreviewGuildResponse>>>;

    fn get_banned_users(
        &self, 
        request: HrpcRequest<GetBannedUsersRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetBannedUsersResponse>>>;

    fn ban_user(
        &self, 
        request: HrpcRequest<BanUserRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<BanUserResponse>>>;

    fn kick_user(
        &self, 
        request: HrpcRequest<KickUserRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<KickUserResponse>>>;

    fn unban_user(
        &self, 
        request: HrpcRequest<UnbanUserRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UnbanUserResponse>>>;

    fn get_pinned_messages(
        &self, 
        request: HrpcRequest<GetPinnedMessagesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPinnedMessagesResponse>>>;

    fn pin_message(
        &self, 
        request: HrpcRequest<PinMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<PinMessageResponse>>>;

    fn unpin_message(
        &self, 
        request: HrpcRequest<UnpinMessageRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<UnpinMessageResponse>>>;

    fn stream_events(
        &self, 
        request: HrpcRequest<()>, 
        socket: Socket<StreamEventsResponse, StreamEventsRequest>
    ) -> BoxFuture<'_, ServerResult<()>>;

    fn add_reaction(
        &self, 
        request: HrpcRequest<AddReactionRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<AddReactionResponse>>>;

    fn remove_reaction(
        &self, 
        request: HrpcRequest<RemoveReactionRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<RemoveReactionResponse>>>;

    fn grant_ownership(
        &self, 
        request: HrpcRequest<GrantOwnershipRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GrantOwnershipResponse>>>;

    fn give_up_ownership(
        &self, 
        request: HrpcRequest<GiveUpOwnershipRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GiveUpOwnershipResponse>>>;

    fn create_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn create_room_middleware(&self) -> Option<HrpcLayer> { ... }

    fn create_direct_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn upgrade_room_to_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn create_invite_middleware(&self) -> Option<HrpcLayer> { ... }

    fn create_channel_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_list_middleware(&self) -> Option<HrpcLayer> { ... }

    fn invite_user_to_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_pending_invites_middleware(&self) -> Option<HrpcLayer> { ... }

    fn reject_pending_invite_middleware(&self) -> Option<HrpcLayer> { ... }

    fn ignore_pending_invite_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_invites_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_members_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_channels_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_channel_messages_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_guild_information_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_channel_information_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_channel_order_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_all_channel_order_middleware(&self) -> Option<HrpcLayer> { ... }

    fn update_message_text_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_invite_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_channel_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn join_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn leave_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn trigger_action_middleware(&self) -> Option<HrpcLayer> { ... }

    fn send_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn query_has_permission_middleware(&self) -> Option<HrpcLayer> { ... }

    fn set_permissions_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_permissions_middleware(&self) -> Option<HrpcLayer> { ... }

    fn move_role_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_guild_roles_middleware(&self) -> Option<HrpcLayer> { ... }

    fn add_guild_role_middleware(&self) -> Option<HrpcLayer> { ... }

    fn modify_guild_role_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_guild_role_middleware(&self) -> Option<HrpcLayer> { ... }

    fn manage_user_roles_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_user_roles_middleware(&self) -> Option<HrpcLayer> { ... }

    fn typing_middleware(&self) -> Option<HrpcLayer> { ... }

    fn preview_guild_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_banned_users_middleware(&self) -> Option<HrpcLayer> { ... }

    fn ban_user_middleware(&self) -> Option<HrpcLayer> { ... }

    fn kick_user_middleware(&self) -> Option<HrpcLayer> { ... }

    fn unban_user_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_pinned_messages_middleware(&self) -> Option<HrpcLayer> { ... }

    fn pin_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn unpin_message_middleware(&self) -> Option<HrpcLayer> { ... }

    fn stream_events_middleware(&self) -> Option<HrpcLayer> { ... }

    fn add_reaction_middleware(&self) -> Option<HrpcLayer> { ... }

    fn remove_reaction_middleware(&self) -> Option<HrpcLayer> { ... }

    fn grant_ownership_middleware(&self) -> Option<HrpcLayer> { ... }

    fn give_up_ownership_middleware(&self) -> Option<HrpcLayer> { ... }

}

Generated trait containing hRPC methods that should be implemented for use with ChatServiceServer.

Required methods

fn create_guild(
&self,
request: HrpcRequest<CreateGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateGuildResponse>>>

Endpoint to create a guild.

fn create_room(
&self,
request: HrpcRequest<CreateRoomRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateRoomResponse>>>

Endpoint to create a “room” guild.

fn create_direct_message(
&self,
request: HrpcRequest<CreateDirectMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateDirectMessageResponse>>>

Endpoint to create a “direct message” guild.

The inviter and the invitee that join the created guild will both be owners.

The guild should be created on the server that inviter is on.

The server should send a guild invite to the invitee (specified in the request), using the UserInvited postbox event if the invitee is on another server, otherwise see the below item.

The server should process this as follows: adding the invite to their pending invite list and sending an InviteReceived event over their event stream if the invitee is on this server.

The invitee may or may not use the invite. Only the invitee may use the invite.

fn upgrade_room_to_guild(
&self,
request: HrpcRequest<UpgradeRoomToGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpgradeRoomToGuildResponse>>>

Endpoint to upgrade a “room” guild to a “normal” guild.

fn create_invite(
&self,
request: HrpcRequest<CreateInviteRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateInviteResponse>>>

Endpoint to create an invite.

fn create_channel(
&self,
request: HrpcRequest<CreateChannelRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateChannelResponse>>>

Endpoint to create a channel.

fn get_guild_list(
&self,
request: HrpcRequest<GetGuildListRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildListResponse>>>

Endpoint to get your guild list.

fn invite_user_to_guild(
&self,
request: HrpcRequest<InviteUserToGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<InviteUserToGuildResponse>>>

Endpoint to invite a user to a guild.

fn get_pending_invites(
&self,
request: HrpcRequest<GetPendingInvitesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPendingInvitesResponse>>>

Endpoint to get your pending invites.

fn reject_pending_invite(
&self,
request: HrpcRequest<RejectPendingInviteRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<RejectPendingInviteResponse>>>

Endpoint to reject (delete with notification to the inviter) an invite from your pending invite list.

If the invitee is on a different server than the inviter, the invitee’s server should send a UserRejectedInvite postbox event to the inviter’s server.

The “notification” is sending a InviteRejected stream event to the inviter. If the guild’s kind is DirectMessage, the server should also set the rejected field of the guild’s kind to true.

fn ignore_pending_invite(
&self,
request: HrpcRequest<IgnorePendingInviteRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<IgnorePendingInviteResponse>>>

Endpoint to ignore (delete without notification to the inviter) an invite from your pending invite list.

fn get_guild(
&self,
request: HrpcRequest<GetGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildResponse>>>

Endpoint to get information about a guild.

fn get_guild_invites(
&self,
request: HrpcRequest<GetGuildInvitesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildInvitesResponse>>>

Endpoint to get the invites of a guild.

This requires the “invites.view” permission.

fn get_guild_members(
&self,
request: HrpcRequest<GetGuildMembersRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildMembersResponse>>>

Endpoint to get the members of a guild.

fn get_guild_channels(
&self,
request: HrpcRequest<GetGuildChannelsRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildChannelsResponse>>>

Endpoint to get the channels of a guild.

You will only be informed of channels you have the “messages.view” permission for.

fn get_channel_messages(
&self,
request: HrpcRequest<GetChannelMessagesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetChannelMessagesResponse>>>

Endpoint to get the messages from a guild channel.

fn get_message(
&self,
request: HrpcRequest<GetMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetMessageResponse>>>

Endpoint to get a single message from a guild channel.

fn update_guild_information(
&self,
request: HrpcRequest<UpdateGuildInformationRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateGuildInformationResponse>>>

Endpoint to update a guild’s information.

fn update_channel_information(
&self,
request: HrpcRequest<UpdateChannelInformationRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateChannelInformationResponse>>>

Endpoint to update a channel’s information.

fn update_channel_order(
&self,
request: HrpcRequest<UpdateChannelOrderRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateChannelOrderResponse>>>

Endpoint to change the position of a channel in the channel list.

fn update_all_channel_order(
&self,
request: HrpcRequest<UpdateAllChannelOrderRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateAllChannelOrderResponse>>>

Endpoint to change the position of all channels in the guild; must specifcy all channels or fails

fn update_message_text(
&self,
request: HrpcRequest<UpdateMessageTextRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UpdateMessageTextResponse>>>

Endpoint to change the text of a message.

fn delete_guild(
&self,
request: HrpcRequest<DeleteGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteGuildResponse>>>

Endpoint to delete a guild. Can only be invoked if there’s one owner.

fn delete_invite(
&self,
request: HrpcRequest<DeleteInviteRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteInviteResponse>>>

Endpoint to delete an invite.

fn delete_channel(
&self,
request: HrpcRequest<DeleteChannelRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteChannelResponse>>>

Endpoint to delete a channel.

fn delete_message(
&self,
request: HrpcRequest<DeleteMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteMessageResponse>>>

Endpoint to delete a message.

This requires the “messages.manage.delete” permission if you are not the message author.

fn join_guild(
&self,
request: HrpcRequest<JoinGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<JoinGuildResponse>>>

Endpoint to join a guild.

If the invite used is in a user’s “pending invites” list, it should be removed from there.

fn leave_guild(
&self,
request: HrpcRequest<LeaveGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<LeaveGuildResponse>>>

Endpoint to leave a guild.

If you’re the only owner, you can’t leave a guild. Exception to this rule are “direct message” guilds.

When the last member in a “direct message” guild leaves it, that guild should be deleted.

fn trigger_action(
&self,
request: HrpcRequest<TriggerActionRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<TriggerActionResponse>>>

Endpont to trigger an action.

fn send_message(
&self,
request: HrpcRequest<SendMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<SendMessageResponse>>>

Endpoint to send a message.

fn query_has_permission(
&self,
request: HrpcRequest<QueryHasPermissionRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<QueryHasPermissionResponse>>>

Endpoint to query if a user has a permission.

fn set_permissions(
&self,
request: HrpcRequest<SetPermissionsRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<SetPermissionsResponse>>>

Endpoint to set permissions for a role in a guild.

fn get_permissions(
&self,
request: HrpcRequest<GetPermissionsRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPermissionsResponse>>>

Endpoint to get permissions for a role in a guild.

fn move_role(
&self,
request: HrpcRequest<MoveRoleRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<MoveRoleResponse>>>

Endpoint to change the position of a role in the role list in a guild.

fn get_guild_roles(
&self,
request: HrpcRequest<GetGuildRolesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetGuildRolesResponse>>>

Endpoint to get the roles from a guild.

fn add_guild_role(
&self,
request: HrpcRequest<AddGuildRoleRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<AddGuildRoleResponse>>>

Endpoint to add a role to a guild.

fn modify_guild_role(
&self,
request: HrpcRequest<ModifyGuildRoleRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<ModifyGuildRoleResponse>>>

Endpoint to a modify a role from a guild.

fn delete_guild_role(
&self,
request: HrpcRequest<DeleteGuildRoleRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteGuildRoleResponse>>>

Endpoint to delete a role from a guild.

fn manage_user_roles(
&self,
request: HrpcRequest<ManageUserRolesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<ManageUserRolesResponse>>>

Endpoint to manage the roles of a guild member.

fn get_user_roles(
&self,
request: HrpcRequest<GetUserRolesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetUserRolesResponse>>>

Endpoint to get the roles a guild member has.

fn typing(
&self,
request: HrpcRequest<TypingRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<TypingResponse>>>

Endpoint to send a typing notification in a guild channel.

fn preview_guild(
&self,
request: HrpcRequest<PreviewGuildRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<PreviewGuildResponse>>>

Endpoint to “preview” a guild, which returns some information about a guild.

Guilds with the “direct message” kind can only be previewed by the user who is invited to the guild.

fn get_banned_users(
&self,
request: HrpcRequest<GetBannedUsersRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetBannedUsersResponse>>>

Endpoint to get banned users in a guild.

fn ban_user(
&self,
request: HrpcRequest<BanUserRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<BanUserResponse>>>

Endpoint to ban a user from a guild.

fn kick_user(
&self,
request: HrpcRequest<KickUserRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<KickUserResponse>>>

Endpoint to kick a user from a guild.

fn unban_user(
&self,
request: HrpcRequest<UnbanUserRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UnbanUserResponse>>>

Endpoint to unban a user from a guild.

fn get_pinned_messages(
&self,
request: HrpcRequest<GetPinnedMessagesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetPinnedMessagesResponse>>>

Endpoint to get all pinned messages in a guild channel.

fn pin_message(
&self,
request: HrpcRequest<PinMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<PinMessageResponse>>>

Endpoint to pin a message in a guild channel.

fn unpin_message(
&self,
request: HrpcRequest<UnpinMessageRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<UnpinMessageResponse>>>

Endpoint to unpin a message in a guild channel.

fn stream_events(
&self,
request: HrpcRequest<()>,
socket: Socket<StreamEventsResponse, StreamEventsRequest>
) -> BoxFuture<'_, ServerResult<()>>

Endpoint to stream events from the homeserver.

fn add_reaction(
&self,
request: HrpcRequest<AddReactionRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<AddReactionResponse>>>

Endpoint to add a reaction to a message.

fn remove_reaction(
&self,
request: HrpcRequest<RemoveReactionRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<RemoveReactionResponse>>>

Endpoint to remove a reaction from a message.

fn grant_ownership(
&self,
request: HrpcRequest<GrantOwnershipRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GrantOwnershipResponse>>>

Endpoint to give ownership to someone else.

fn give_up_ownership(
&self,
request: HrpcRequest<GiveUpOwnershipRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GiveUpOwnershipResponse>>>

Endpoint to give up your own ownership. Requires that at least one other person will still be owner.

Provided methods

fn create_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn create_room_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn create_direct_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn upgrade_room_to_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn create_invite_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn create_channel_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_list_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn invite_user_to_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_pending_invites_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn reject_pending_invite_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn ignore_pending_invite_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_invites_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_members_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_channels_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_channel_messages_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_guild_information_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_channel_information_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_channel_order_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_all_channel_order_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn update_message_text_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_invite_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_channel_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn join_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn leave_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn trigger_action_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn send_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn query_has_permission_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn set_permissions_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_permissions_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn move_role_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_guild_roles_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn add_guild_role_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn modify_guild_role_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_guild_role_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn manage_user_roles_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_user_roles_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn typing_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn preview_guild_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_banned_users_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn ban_user_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn kick_user_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn unban_user_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_pinned_messages_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn pin_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn unpin_message_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn stream_events_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn add_reaction_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn remove_reaction_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn grant_ownership_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn give_up_ownership_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.