Trait harmony_rust_sdk::api::chat::v1::chat_service_server::ChatService

pub trait ChatService: Send + Sync + 'static {
    type Error: CustomError + Send + Sync + 'static;
    type StreamEventsValidationType: Send + Sync;

    #[must_use]
    fn create_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<CreateGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn create_invite<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateInviteRequest>
    ) -> Pin<Box<dyn Future<Output = Result<CreateInviteResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn create_channel<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateChannelRequest>
    ) -> Pin<Box<dyn Future<Output = Result<CreateChannelResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild_list<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildListRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildListResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild_invites<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildInvitesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildInvitesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild_members<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildMembersRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildMembersResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild_channels<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildChannelsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildChannelsResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_channel_messages<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetChannelMessagesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetChannelMessagesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetMessageResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_guild_information<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateGuildInformationRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateGuildInformationResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_channel_information<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateChannelInformationRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateChannelInformationResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_channel_order<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateChannelOrderRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateChannelOrderResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_all_channel_order<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateAllChannelOrderRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateAllChannelOrderResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_message_text<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateMessageTextRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateMessageTextResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_invite<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteInviteRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteInviteResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_channel<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteChannelRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteChannelResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteMessageResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn join_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<JoinGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<JoinGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn leave_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<LeaveGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<LeaveGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn trigger_action<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<TriggerActionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<TriggerActionResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn send_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SendMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<SendMessageResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn query_has_permission<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<QueryHasPermissionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<QueryHasPermissionResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn set_permissions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SetPermissionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<SetPermissionsResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_permissions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetPermissionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetPermissionsResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn move_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<MoveRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<MoveRoleResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_guild_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetGuildRolesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn add_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<AddGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<AddGuildRoleResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn modify_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ModifyGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<ModifyGuildRoleResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn delete_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<DeleteGuildRoleResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn manage_user_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ManageUserRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<ManageUserRolesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_user_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetUserRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetUserRolesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn typing<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<TypingRequest>
    ) -> Pin<Box<dyn Future<Output = Result<TypingResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn preview_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<PreviewGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<PreviewGuildResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn ban_user<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<BanUserRequest>
    ) -> Pin<Box<dyn Future<Output = Result<BanUserResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn kick_user<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<KickUserRequest>
    ) -> Pin<Box<dyn Future<Output = Result<KickUserResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn unban_user<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UnbanUserRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UnbanUserResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_pinned_messages<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetPinnedMessagesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetPinnedMessagesResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn pin_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<PinMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<PinMessageResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn unpin_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UnpinMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UnpinMessageResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn stream_events_validation<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Option<StreamEventsRequest>>
    ) -> Pin<Box<dyn Future<Output = Result<Self::StreamEventsValidationType, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn stream_events<'life0, 'async_trait>(
        &'life0 self, 
        validation_value: Self::StreamEventsValidationType, 
        socket: Socket<StreamEventsRequest, StreamEventsResponse>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn create_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn create_invite_pre(&self) -> BoxedFilter<()> { ... }

    fn create_channel_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_list_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_invites_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_members_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_channels_pre(&self) -> BoxedFilter<()> { ... }

    fn get_channel_messages_pre(&self) -> BoxedFilter<()> { ... }

    fn get_message_pre(&self) -> BoxedFilter<()> { ... }

    fn update_guild_information_pre(&self) -> BoxedFilter<()> { ... }

    fn update_channel_information_pre(&self) -> BoxedFilter<()> { ... }

    fn update_channel_order_pre(&self) -> BoxedFilter<()> { ... }

    fn update_all_channel_order_pre(&self) -> BoxedFilter<()> { ... }

    fn update_message_text_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_invite_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_channel_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_message_pre(&self) -> BoxedFilter<()> { ... }

    fn join_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn leave_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn trigger_action_pre(&self) -> BoxedFilter<()> { ... }

    fn send_message_pre(&self) -> BoxedFilter<()> { ... }

    fn query_has_permission_pre(&self) -> BoxedFilter<()> { ... }

    fn set_permissions_pre(&self) -> BoxedFilter<()> { ... }

    fn get_permissions_pre(&self) -> BoxedFilter<()> { ... }

    fn move_role_pre(&self) -> BoxedFilter<()> { ... }

    fn get_guild_roles_pre(&self) -> BoxedFilter<()> { ... }

    fn add_guild_role_pre(&self) -> BoxedFilter<()> { ... }

    fn modify_guild_role_pre(&self) -> BoxedFilter<()> { ... }

    fn delete_guild_role_pre(&self) -> BoxedFilter<()> { ... }

    fn manage_user_roles_pre(&self) -> BoxedFilter<()> { ... }

    fn get_user_roles_pre(&self) -> BoxedFilter<()> { ... }

    fn typing_pre(&self) -> BoxedFilter<()> { ... }

    fn preview_guild_pre(&self) -> BoxedFilter<()> { ... }

    fn ban_user_pre(&self) -> BoxedFilter<()> { ... }

    fn kick_user_pre(&self) -> BoxedFilter<()> { ... }

    fn unban_user_pre(&self) -> BoxedFilter<()> { ... }

    fn get_pinned_messages_pre(&self) -> BoxedFilter<()> { ... }

    fn pin_message_pre(&self) -> BoxedFilter<()> { ... }

    fn unpin_message_pre(&self) -> BoxedFilter<()> { ... }

    fn stream_events_pre(&self) -> BoxedFilter<()> { ... }

    fn stream_events_on_upgrade(&self, response: Response) -> Response { ... }

}

Generated trait containing hRPC methods that should be implemented for use with ChatServiceServer.

Associated Types

type Error: CustomError + Send + Sync + 'static

type StreamEventsValidationType: Send + Sync

Loading content...
Required methods

#[must_use]
fn create_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<CreateGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to create a guild.

#[must_use]
fn create_invite<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateInviteRequest>
) -> Pin<Box<dyn Future<Output = Result<CreateInviteResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to create an invite.

#[must_use]
fn create_channel<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateChannelRequest>
) -> Pin<Box<dyn Future<Output = Result<CreateChannelResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to create a channel.

#[must_use]
fn get_guild_list<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildListRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildListResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get your guild list.

#[must_use]
fn get_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get information about a guild.

#[must_use]
fn get_guild_invites<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildInvitesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildInvitesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the invites of a guild.

This requires the “invites.view” permission.

#[must_use]
fn get_guild_members<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildMembersRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildMembersResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the members of a guild.

#[must_use]
fn get_guild_channels<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildChannelsRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildChannelsResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the channels of a guild.

You will only be informed of channels you have the “messages.view” permission for.

#[must_use]
fn get_channel_messages<'life0, 'async_trait>(
&'life0 self,
request: Request<GetChannelMessagesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetChannelMessagesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the messages from a guild channel.

#[must_use]
fn get_message<'life0, 'async_trait>(
&'life0 self,
request: Request<GetMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<GetMessageResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get a single message from a guild channel.

#[must_use]
fn update_guild_information<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateGuildInformationRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateGuildInformationResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to update a guild’s information.

#[must_use]
fn update_channel_information<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateChannelInformationRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateChannelInformationResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to update a channel’s information.

#[must_use]
fn update_channel_order<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateChannelOrderRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateChannelOrderResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to change the position of a channel in the channel list.

#[must_use]
fn update_all_channel_order<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateAllChannelOrderRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateAllChannelOrderResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to change the position of all channels in the guild; must specifcy all channels or fails

#[must_use]
fn update_message_text<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateMessageTextRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateMessageTextResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to change the text of a message.

#[must_use]
fn delete_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete a guild.

#[must_use]
fn delete_invite<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteInviteRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteInviteResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete an invite.

#[must_use]
fn delete_channel<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteChannelRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteChannelResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete a channel.

#[must_use]
fn delete_message<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteMessageResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete a message.

This requires the “messages.manage.delete” permission if you are not the message author.

#[must_use]
fn join_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<JoinGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<JoinGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to join a guild.

#[must_use]
fn leave_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<LeaveGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<LeaveGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to leave a guild.

#[must_use]
fn trigger_action<'life0, 'async_trait>(
&'life0 self,
request: Request<TriggerActionRequest>
) -> Pin<Box<dyn Future<Output = Result<TriggerActionResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpont to trigger an action.

#[must_use]
fn send_message<'life0, 'async_trait>(
&'life0 self,
request: Request<SendMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<SendMessageResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to send a message.

#[must_use]
fn query_has_permission<'life0, 'async_trait>(
&'life0 self,
request: Request<QueryHasPermissionRequest>
) -> Pin<Box<dyn Future<Output = Result<QueryHasPermissionResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to query if a user has a permission.

#[must_use]
fn set_permissions<'life0, 'async_trait>(
&'life0 self,
request: Request<SetPermissionsRequest>
) -> Pin<Box<dyn Future<Output = Result<SetPermissionsResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to set permissions for a role in a guild.

#[must_use]
fn get_permissions<'life0, 'async_trait>(
&'life0 self,
request: Request<GetPermissionsRequest>
) -> Pin<Box<dyn Future<Output = Result<GetPermissionsResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get permissions for a role in a guild.

#[must_use]
fn move_role<'life0, 'async_trait>(
&'life0 self,
request: Request<MoveRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<MoveRoleResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to change the position of a role in the role list in a guild.

#[must_use]
fn get_guild_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetGuildRolesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the roles from a guild.

#[must_use]
fn add_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<AddGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<AddGuildRoleResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to add a role to a guild.

#[must_use]
fn modify_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<ModifyGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<ModifyGuildRoleResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to a modify a role from a guild.

#[must_use]
fn delete_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<DeleteGuildRoleResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to delete a role from a guild.

#[must_use]
fn manage_user_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<ManageUserRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<ManageUserRolesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to manage the roles of a guild member.

#[must_use]
fn get_user_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<GetUserRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetUserRolesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get the roles a guild member has.

#[must_use]
fn typing<'life0, 'async_trait>(
&'life0 self,
request: Request<TypingRequest>
) -> Pin<Box<dyn Future<Output = Result<TypingResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to send a typing notification in a guild channel.

#[must_use]
fn preview_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<PreviewGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<PreviewGuildResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to “preview” a guild, which returns some information about a guild.

#[must_use]
fn ban_user<'life0, 'async_trait>(
&'life0 self,
request: Request<BanUserRequest>
) -> Pin<Box<dyn Future<Output = Result<BanUserResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to ban a user from a guild.

#[must_use]
fn kick_user<'life0, 'async_trait>(
&'life0 self,
request: Request<KickUserRequest>
) -> Pin<Box<dyn Future<Output = Result<KickUserResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to kick a user from a guild.

#[must_use]
fn unban_user<'life0, 'async_trait>(
&'life0 self,
request: Request<UnbanUserRequest>
) -> Pin<Box<dyn Future<Output = Result<UnbanUserResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to unban a user from a guild.

#[must_use]
fn get_pinned_messages<'life0, 'async_trait>(
&'life0 self,
request: Request<GetPinnedMessagesRequest>
) -> Pin<Box<dyn Future<Output = Result<GetPinnedMessagesResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to get all pinned messages in a guild channel.

#[must_use]
fn pin_message<'life0, 'async_trait>(
&'life0 self,
request: Request<PinMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<PinMessageResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to pin a message in a guild channel.

#[must_use]
fn unpin_message<'life0, 'async_trait>(
&'life0 self,
request: Request<UnpinMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<UnpinMessageResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to unpin a message in a guild channel.

#[must_use]
fn stream_events_validation<'life0, 'async_trait>(
&'life0 self,
request: Request<Option<StreamEventsRequest>>
) -> Pin<Box<dyn Future<Output = Result<Self::StreamEventsValidationType, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]
fn stream_events<'life0, 'async_trait>(
&'life0 self,
validation_value: Self::StreamEventsValidationType,
socket: Socket<StreamEventsRequest, StreamEventsResponse>
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Endpoint to stream events from the homeserver.

Loading content...
Provided methods

fn create_guild_pre(&self) -> BoxedFilter<()>

fn create_invite_pre(&self) -> BoxedFilter<()>

fn create_channel_pre(&self) -> BoxedFilter<()>

fn get_guild_list_pre(&self) -> BoxedFilter<()>

fn get_guild_pre(&self) -> BoxedFilter<()>

fn get_guild_invites_pre(&self) -> BoxedFilter<()>

fn get_guild_members_pre(&self) -> BoxedFilter<()>

fn get_guild_channels_pre(&self) -> BoxedFilter<()>

fn get_channel_messages_pre(&self) -> BoxedFilter<()>

fn get_message_pre(&self) -> BoxedFilter<()>

fn update_guild_information_pre(&self) -> BoxedFilter<()>

fn update_channel_information_pre(&self) -> BoxedFilter<()>

fn update_channel_order_pre(&self) -> BoxedFilter<()>

fn update_all_channel_order_pre(&self) -> BoxedFilter<()>

fn update_message_text_pre(&self) -> BoxedFilter<()>

fn delete_guild_pre(&self) -> BoxedFilter<()>

fn delete_invite_pre(&self) -> BoxedFilter<()>

fn delete_channel_pre(&self) -> BoxedFilter<()>

fn delete_message_pre(&self) -> BoxedFilter<()>

fn join_guild_pre(&self) -> BoxedFilter<()>

fn leave_guild_pre(&self) -> BoxedFilter<()>

fn trigger_action_pre(&self) -> BoxedFilter<()>

fn send_message_pre(&self) -> BoxedFilter<()>

fn query_has_permission_pre(&self) -> BoxedFilter<()>

fn set_permissions_pre(&self) -> BoxedFilter<()>

fn get_permissions_pre(&self) -> BoxedFilter<()>

fn move_role_pre(&self) -> BoxedFilter<()>

fn get_guild_roles_pre(&self) -> BoxedFilter<()>

fn add_guild_role_pre(&self) -> BoxedFilter<()>

fn modify_guild_role_pre(&self) -> BoxedFilter<()>

fn delete_guild_role_pre(&self) -> BoxedFilter<()>

fn manage_user_roles_pre(&self) -> BoxedFilter<()>

fn get_user_roles_pre(&self) -> BoxedFilter<()>

fn typing_pre(&self) -> BoxedFilter<()>

fn preview_guild_pre(&self) -> BoxedFilter<()>

fn ban_user_pre(&self) -> BoxedFilter<()>

fn kick_user_pre(&self) -> BoxedFilter<()>

fn unban_user_pre(&self) -> BoxedFilter<()>

fn get_pinned_messages_pre(&self) -> BoxedFilter<()>

fn pin_message_pre(&self) -> BoxedFilter<()>

fn unpin_message_pre(&self) -> BoxedFilter<()>

fn stream_events_pre(&self) -> BoxedFilter<()>

fn stream_events_on_upgrade(&self, response: Response) -> Response

Loading content...