Trait harmony_rust_sdk::api::chat::v1::chat_service_server::ChatService

pub trait ChatService: Send + Sync + 'static {
    type StreamEventsStream: Stream<Item = Result<Event, Status>> + Send + Sync + 'static;
    type SyncStream: Stream<Item = Result<SyncEvent, Status>> + Send + Sync + 'static;
#[must_use]    fn create_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateGuildResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn create_invite<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateInviteRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateInviteResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn create_channel<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateChannelRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateChannelResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn create_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<CreateEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<CreateEmotePackResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild_list<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildListRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildListResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn add_guild_to_guild_list<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<AddGuildToGuildListRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<AddGuildToGuildListResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn remove_guild_from_guild_list<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<RemoveGuildFromGuildListRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<RemoveGuildFromGuildListResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild_invites<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildInvitesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildInvitesResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild_members<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildMembersRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildMembersResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild_channels<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildChannelsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildChannelsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_channel_messages<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetChannelMessagesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetChannelMessagesResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetMessageResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_emote_packs<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetEmotePacksRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetEmotePacksResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_emote_pack_emotes<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetEmotePackEmotesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetEmotePackEmotesResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update_guild_information<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateGuildInformationRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update_channel_information<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateChannelInformationRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update_channel_order<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateChannelOrderRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn update_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn add_emote_to_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<AddEmoteToPackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_invite<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteInviteRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_channel<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteChannelRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_emote_from_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteEmoteFromPackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn dequip_emote_pack<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DequipEmotePackRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn join_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<JoinGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<JoinGuildResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn leave_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<LeaveGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn trigger_action<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<TriggerActionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn send_message<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SendMessageRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<SendMessageResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn query_has_permission<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<QueryPermissionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<QueryPermissionsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn set_permissions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SetPermissionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_permissions<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetPermissionsRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetPermissionsResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn move_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<MoveRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<MoveRoleResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_guild_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetGuildRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildRolesResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn add_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<AddGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<AddGuildRoleResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn modify_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ModifyGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn delete_guild_role<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<DeleteGuildRoleRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn manage_user_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ManageUserRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_user_roles<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetUserRolesRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetUserRolesResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn stream_events<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<Streaming<StreamEventsRequest>>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamEventsStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn sync<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SyncRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SyncStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_user<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetUserRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetUserResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn get_user_metadata<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetUserMetadataRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<GetUserMetadataResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn profile_update<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<ProfileUpdateRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn typing<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<TypingRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

#[must_use]    fn preview_guild<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<PreviewGuildRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<PreviewGuildResponse>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ChatServiceServer.

Associated Types

type StreamEventsStream: Stream<Item = Result<Event, Status>> + Send + Sync + 'static

Server streaming response type for the StreamEvents method.

type SyncStream: Stream<Item = Result<SyncEvent, Status>> + Send + Sync + 'static

Server streaming response type for the Sync method.

Loading content...
Required methods

#[must_use]fn create_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<CreateGuildResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn create_invite<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateInviteRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<CreateInviteResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn create_channel<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateChannelRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<CreateChannelResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn create_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<CreateEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<CreateEmotePackResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_guild_list<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildListRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildListResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn add_guild_to_guild_list<'life0, 'async_trait>(
&'life0 self,
request: Request<AddGuildToGuildListRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<AddGuildToGuildListResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn remove_guild_from_guild_list<'life0, 'async_trait>(
&'life0 self,
request: Request<RemoveGuildFromGuildListRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<RemoveGuildFromGuildListResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_guild_invites<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildInvitesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildInvitesResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This requires the "invites.view" permission.

#[must_use]fn get_guild_members<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildMembersRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildMembersResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_guild_channels<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildChannelsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildChannelsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

You will only be informed of channels you have the "messages.view" permission for.

#[must_use]fn get_channel_messages<'life0, 'async_trait>(
&'life0 self,
request: Request<GetChannelMessagesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetChannelMessagesResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_message<'life0, 'async_trait>(
&'life0 self,
request: Request<GetMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetMessageResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_emote_packs<'life0, 'async_trait>(
&'life0 self,
request: Request<GetEmotePacksRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetEmotePacksResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_emote_pack_emotes<'life0, 'async_trait>(
&'life0 self,
request: Request<GetEmotePackEmotesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetEmotePackEmotesResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn update_guild_information<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateGuildInformationRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn update_channel_information<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateChannelInformationRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn update_channel_order<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateChannelOrderRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn update_message<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn add_emote_to_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<AddEmoteToPackRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_invite<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteInviteRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_channel<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteChannelRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_message<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This requires the "messages.manage.delete" permission if you are not the message author.

#[must_use]fn delete_emote_from_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteEmoteFromPackRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn dequip_emote_pack<'life0, 'async_trait>(
&'life0 self,
request: Request<DequipEmotePackRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn join_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<JoinGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<JoinGuildResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn leave_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<LeaveGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn trigger_action<'life0, 'async_trait>(
&'life0 self,
request: Request<TriggerActionRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn send_message<'life0, 'async_trait>(
&'life0 self,
request: Request<SendMessageRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<SendMessageResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn query_has_permission<'life0, 'async_trait>(
&'life0 self,
request: Request<QueryPermissionsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<QueryPermissionsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn set_permissions<'life0, 'async_trait>(
&'life0 self,
request: Request<SetPermissionsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_permissions<'life0, 'async_trait>(
&'life0 self,
request: Request<GetPermissionsRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetPermissionsResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn move_role<'life0, 'async_trait>(
&'life0 self,
request: Request<MoveRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<MoveRoleResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_guild_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<GetGuildRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetGuildRolesResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn add_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<AddGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<AddGuildRoleResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn modify_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<ModifyGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn delete_guild_role<'life0, 'async_trait>(
&'life0 self,
request: Request<DeleteGuildRoleRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn manage_user_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<ManageUserRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_user_roles<'life0, 'async_trait>(
&'life0 self,
request: Request<GetUserRolesRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetUserRolesResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn stream_events<'life0, 'async_trait>(
&'life0 self,
request: Request<Streaming<StreamEventsRequest>>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::StreamEventsStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn sync<'life0, 'async_trait>(
&'life0 self,
request: Request<SyncRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::SyncStream>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_user<'life0, 'async_trait>(
&'life0 self,
request: Request<GetUserRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetUserResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn get_user_metadata<'life0, 'async_trait>(
&'life0 self,
request: Request<GetUserMetadataRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<GetUserMetadataResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn profile_update<'life0, 'async_trait>(
&'life0 self,
request: Request<ProfileUpdateRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn typing<'life0, 'async_trait>(
&'life0 self,
request: Request<TypingRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

#[must_use]fn preview_guild<'life0, 'async_trait>(
&'life0 self,
request: Request<PreviewGuildRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<PreviewGuildResponse>, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Loading content...