Module harmony_rust_sdk::api::chat::v1::format

Nested message and enum types in Format.

Modules

color

Nested message and enum types in Color.

Structs

ArchivedBold

An archived Bold

ArchivedChannelMention

An archived ChannelMention

ArchivedCodeBlock

An archived CodeBlock

ArchivedColor

An archived Color

ArchivedEmoji

An archived Emoji

ArchivedGuildMention

An archived GuildMention

ArchivedItalic

An archived Italic

ArchivedLocalization

An archived Localization

ArchivedMonospace

An archived Monospace

ArchivedRoleMention

An archived RoleMention

ArchivedSubscript

An archived Subscript

ArchivedSuperscript

An archived Superscript

ArchivedUnderline

An archived Underline

ArchivedUserMention

An archived UserMention

Bold

Bold text

BoldResolver

The resolver for an archived Bold

ChannelMention

Mention of a channel (on the current guild)

ChannelMentionResolver

The resolver for an archived ChannelMention

CodeBlock

A larger codeblock, with a programming language specified Clients should ideally not bound the width of codeblock messages, possibly scrolling the codeblock horizontally separately from the rest of the message

CodeBlockResolver

The resolver for an archived CodeBlock

Color

Colour modification

ColorResolver

The resolver for an archived Color

Emoji

An emoji

EmojiResolver

The resolver for an archived Emoji

GuildMention

Mention of a guild

GuildMentionResolver

The resolver for an archived GuildMention

Italic

Italic text

ItalicResolver

The resolver for an archived Italic

Localization

Replace a part of the text with the text matching the i18n code. If i18n code was not found, keep the original text.

LocalizationResolver

The resolver for an archived Localization

Monospace

Monospace text

MonospaceResolver

The resolver for an archived Monospace

RoleMention

Mention of a role (on the current guild)

RoleMentionResolver

The resolver for an archived RoleMention

Subscript

Subscript text

SubscriptResolver

The resolver for an archived Subscript

Superscript

Superscript text

SuperscriptResolver

The resolver for an archived Superscript

Underline

Underlined text

UnderlineResolver

The resolver for an archived Underline

UserMention

Mention of a user (on the current homeserver)

UserMentionResolver

The resolver for an archived UserMention

Enums

ArchivedFormat

An archived Format

Format

the style if format to apply to this text

FormatResolver

The resolver for an archived Format