Module harmony_rust_sdk::api::chat::v1::stream_event

Nested message and enum types in StreamEvent.

Structs

ActionPerformed
Event sent when an action is performed.

ChannelCreated
Event sent when a new channel is created.

ChannelDeleted
Event sent when a channel is deleted.

ChannelPositionUpdated
Event sent when a channel’s position in the channel list is changed.

ChannelUpdated
Event sent when a channel’s information is changed.

ChannelsReordered
Event sent when all channels have been reordered

GuildAddedToList
Event sent when you join a new guild.

GuildDeleted
Event sent when a guild is deleted.

GuildRemovedFromList
Event sent when you leave a guild.

GuildUpdated
Event sent when a guild’s information is changed.

MemberJoined
Event sent a user joins to a guild.

MemberLeft
Event sent when a member of a guild leaves said guild for whatever reason.

MessageDeleted
Event sent when a message is deleted.

MessagePinned
Sent when a message is pinned in a guild channel.

MessageSent
Event sent when a new message is sent.

MessageUnpinned
Sent when a message is unpinned in a guild channel.

MessageUpdated
Event sent when a message’s text content is updated.

PermissionUpdated
Event sent when a permission is changed that matters to you.

RoleCreated
Event sent when a role is created.

RoleDeleted
Event sent when a role is deleted.

RoleMoved
Event sent when a role’s position in the role list is changed.

RolePermissionsUpdated
Event sent when a role’s permissions are changed.

RoleUpdated
Event sent when a role’s information is changed.

Typing
Event sent when a user sends a typing notification in a guild channel.

UserRolesUpdated
Event sent when a user’s roles are changed.

Enums

Event
Which event to send.

LeaveReason
A reason for why a user has left a guild.