Trait harmony_rust_sdk::api::emote::v1::emote_service_server::EmoteService

pub trait EmoteService: Sync + Send + 'static {

    fn create_emote_pack(
        &self, 
        request: HrpcRequest<CreateEmotePackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateEmotePackResponse>>>;

    fn get_emote_packs(
        &self, 
        request: HrpcRequest<GetEmotePacksRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetEmotePacksResponse>>>;

    fn get_emote_pack_emotes(
        &self, 
        request: HrpcRequest<GetEmotePackEmotesRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<GetEmotePackEmotesResponse>>>;

    fn add_emote_to_pack(
        &self, 
        request: HrpcRequest<AddEmoteToPackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<AddEmoteToPackResponse>>>;

    fn delete_emote_pack(
        &self, 
        request: HrpcRequest<DeleteEmotePackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteEmotePackResponse>>>;

    fn delete_emote_from_pack(
        &self, 
        request: HrpcRequest<DeleteEmoteFromPackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteEmoteFromPackResponse>>>;

    fn dequip_emote_pack(
        &self, 
        request: HrpcRequest<DequipEmotePackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<DequipEmotePackResponse>>>;

    fn equip_emote_pack(
        &self, 
        request: HrpcRequest<EquipEmotePackRequest>
    ) -> BoxFuture<'_, ServerResult<HrpcResponse<EquipEmotePackResponse>>>;

    fn create_emote_pack_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_emote_packs_middleware(&self) -> Option<HrpcLayer> { ... }

    fn get_emote_pack_emotes_middleware(&self) -> Option<HrpcLayer> { ... }

    fn add_emote_to_pack_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_emote_pack_middleware(&self) -> Option<HrpcLayer> { ... }

    fn delete_emote_from_pack_middleware(&self) -> Option<HrpcLayer> { ... }

    fn dequip_emote_pack_middleware(&self) -> Option<HrpcLayer> { ... }

    fn equip_emote_pack_middleware(&self) -> Option<HrpcLayer> { ... }

}

Generated trait containing hRPC methods that should be implemented for use with EmoteServiceServer.

Required methods

fn create_emote_pack(
&self,
request: HrpcRequest<CreateEmotePackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<CreateEmotePackResponse>>>

Endpoint to create an emote pack.

fn get_emote_packs(
&self,
request: HrpcRequest<GetEmotePacksRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetEmotePacksResponse>>>

Endpoint to get the emote packs you have equipped.

fn get_emote_pack_emotes(
&self,
request: HrpcRequest<GetEmotePackEmotesRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<GetEmotePackEmotesResponse>>>

Endpoint to get the emotes in an emote pack.

fn add_emote_to_pack(
&self,
request: HrpcRequest<AddEmoteToPackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<AddEmoteToPackResponse>>>

Endpoint to add an emote to an emote pack that you own.

fn delete_emote_pack(
&self,
request: HrpcRequest<DeleteEmotePackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteEmotePackResponse>>>

Endpoint to delete an emote pack that you own.

fn delete_emote_from_pack(
&self,
request: HrpcRequest<DeleteEmoteFromPackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DeleteEmoteFromPackResponse>>>

Endpoint to delete an emote from an emote pack.

fn dequip_emote_pack(
&self,
request: HrpcRequest<DequipEmotePackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<DequipEmotePackResponse>>>

Endpoint to dequip an emote pack that you have equipped.

fn equip_emote_pack(
&self,
request: HrpcRequest<EquipEmotePackRequest>
) -> BoxFuture<'_, ServerResult<HrpcResponse<EquipEmotePackResponse>>>

Endpoint to equip an emote pack.

Provided methods

fn create_emote_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_emote_packs_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn get_emote_pack_emotes_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn add_emote_to_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_emote_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn delete_emote_from_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn dequip_emote_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.

fn equip_emote_pack_middleware(&self) -> Option<HrpcLayer>

Optional middleware for this RPC.