Trait harmony_rust_sdk::api::profile::v1::profile_service_server::ProfileService

pub trait ProfileService: Send + Sync + 'static {
    type Error: CustomError + Send + Sync + 'static;
    #[must_use]
    fn get_profile<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetProfileRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetProfileResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn update_profile<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<UpdateProfileRequest>
    ) -> Pin<Box<dyn Future<Output = Result<UpdateProfileResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn get_app_data<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<GetAppDataRequest>
    ) -> Pin<Box<dyn Future<Output = Result<GetAppDataResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    #[must_use]
    fn set_app_data<'life0, 'async_trait>(
        &'life0 self, 
        request: Request<SetAppDataRequest>
    ) -> Pin<Box<dyn Future<Output = Result<SetAppDataResponse, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn get_profile_pre(&self) -> BoxedFilter<()> { ... }

    fn update_profile_pre(&self) -> BoxedFilter<()> { ... }

    fn get_app_data_pre(&self) -> BoxedFilter<()> { ... }

    fn set_app_data_pre(&self) -> BoxedFilter<()> { ... }
}

Generated trait containing hRPC methods that should be implemented for use with ProfileServiceServer.

Associated Types

type Error: CustomError + Send + Sync + 'static

Loading content...
Required methods

#[must_use]
fn get_profile<'life0, 'async_trait>(
&'life0 self,
request: Request<GetProfileRequest>
) -> Pin<Box<dyn Future<Output = Result<GetProfileResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets a user’s profile.

#[must_use]
fn update_profile<'life0, 'async_trait>(
&'life0 self,
request: Request<UpdateProfileRequest>
) -> Pin<Box<dyn Future<Output = Result<UpdateProfileResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Updates the user’s profile.

#[must_use]
fn get_app_data<'life0, 'async_trait>(
&'life0 self,
request: Request<GetAppDataRequest>
) -> Pin<Box<dyn Future<Output = Result<GetAppDataResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Gets app data for a user (this can be used to store user preferences which is synchronized across devices).

#[must_use]
fn set_app_data<'life0, 'async_trait>(
&'life0 self,
request: Request<SetAppDataRequest>
) -> Pin<Box<dyn Future<Output = Result<SetAppDataResponse, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Sets the app data for a user.

Loading content...
Provided methods

fn get_profile_pre(&self) -> BoxedFilter<()>

fn update_profile_pre(&self) -> BoxedFilter<()>

fn get_app_data_pre(&self) -> BoxedFilter<()>

fn set_app_data_pre(&self) -> BoxedFilter<()>

Loading content...