Module harmony_rust_sdk::client

Rust client implementation for Harmony, powered by tonic.

See the examples directory in the repository on how to use this.

Re-exports

pub use crate::api::auth::Session;

pub use error::*;

Modules

api
Client API implementations.

error
Error related code used by Client.

exports
Some crates exported for user convenience.

Structs

Client
Client implementation for Harmony.

Enums

AuthStatus
Represents an authentication state in which a Client can be.