Module harmony_rust_sdk::client::api

Client API implementations.

Modules

auth
Auth service client API.

batch
Batch client API.

chat
Chat service client API.

emote
Emote client API.

harmonytypes
Common types used in other services.

mediaproxy
Media proxy service client API.

profile
Profile client API.

rest
REST client API.

Structs

Hmc
A HMC.

Enums

HmcFromStrError
Errors that can occur when converting a possibly non-HMC string to a HMC URL.

HmcParseError
Errors that can occur when converting a possibly non-HMC URL to a HMC URL.