Struct harmony_rust_sdk::client::Client

pub struct Client { /* fields omitted */ }

Client implementation for Harmony.

Implementations

impl Client

pub async fn new(
homeserver_url: Uri,
session: Option<Session>
) -> ClientResult<Self>

Create a new Client from a homeserver Uri (URL) and an (optional) session.

If port is not specified in the URL, this will add the default port 2289 to it. If scheme is not specified, this will assume the scheme is https.

Example

let client = Client::new("https://example.org".parse().unwrap(), None).await?;

pub fn auth_status(&self) -> AuthStatus

Get the current auth status.

pub fn homeserver_url(&self) -> &Uri

Get the stored homeserver URL.

pub async fn begin_auth<'_>(&'_ self) -> ClientResult<()>

Start an authentication session.

pub async fn next_auth_step<'_>(
&'_ self,
response: AuthStepResponse
) -> ClientResult<Option<AuthStep>>

Request the next authentication step from the server.

Returns Ok(None) if authentication was completed. Returns Ok(Some(AuthStep)) if extra step is requested from the server.

pub async fn prev_auth_step<'_>(&'_ self) -> ClientResult<AuthStep>

Go back to the previous authentication step.

pub async fn auth_stream<'_>(
&'_ self
) -> ClientResult<impl Stream<Item = ClientResult<AuthStep>> + Send + Sync>

Begin an authentication steps stream for the current authentication session.

pub async fn subscribe_events<'_>(
&'_ self,
subscriptions: Vec<EventSource>
) -> ClientResult<impl Stream<Item = ClientResult<Event>> + Send + Sync>

Subscribe to events coming from specified event sources.