Struct harmony_rust_sdk::client::Client

pub struct Client { /* fields omitted */ }

Client implementation for Harmony.

Implementations

impl Client

pub async fn new(
homeserver_url: Uri,
session: Option<Session>
) -> ClientResult<Self>

Create a new Client from a homeserver Uri and an (optional) session.

If port is not specified in the URL, this will add the default port 2289 to it. If scheme is not specified, this will assume the scheme is https.

Example

let client = Client::new("https://chat.harmonyapp.io:2289".parse().unwrap(), None).await?;

pub async fn event_loop<'a, Fut, Hndlr>(
&'a self,
subs: Vec<EventSource>,
handler: Hndlr
) -> Result<(), ClientError> where
Fut: Future<Output = ClientResult<bool>>,
Hndlr: FnMut(&'a Client, Event) -> Fut + Send + 'a,

Starts an event loop with the given handler.

All socket errors will be logged with tracing. If the handler function returns Ok(true) or Err(err), the function will return, so if you don’t want it to return, return Ok(false).

pub fn chat(
&self
) -> MutexGuard<'_, ChatServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the chat service.

pub fn auth(
&self
) -> MutexGuard<'_, AuthServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the auth service.

pub fn mediaproxy(
&self
) -> MutexGuard<'_, MediaProxyServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the mediaproxy service.

pub fn profile(
&self
) -> MutexGuard<'_, ProfileServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the profile service.

pub fn emote(
&self
) -> MutexGuard<'_, EmoteServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the emote service.

pub fn batch(
&self
) -> MutexGuard<'_, BatchServiceClient<Trace<AddAuth<Hyper>, fn(_: &BoxRequest) -> Span, fn(_: &BoxRequest, _: &Span), fn(_: &BoxResponse, _: &Span), fn(_: &BoxResponse, _: &Span, _: &HrpcError)>>>

Get a mutex guard to the batch service.

pub fn call<Req>(
&self,
request: Req
) -> impl Future<Output = ClientResult<Req::Response>> + Send + 'static where
Req: Endpoint,
Req::Response: Message + Default + 'static,

Execute the given request.

pub fn call_response<Req>(
&self,
request: Req
) -> impl Future<Output = ClientResult<Response<Req::Response>>> + Send + 'static where
Req: Endpoint,
Req::Response: 'static,

Execute the given request, but return a [hrpc::Response].

pub fn batch_call<Req>(
&self,
requests: Vec<Req>
) -> impl Future<Output = ClientResult<Vec<<Req as Endpoint>::Response>>> + Send + 'static where
Req: Endpoint + Message,
<Req as Endpoint>::Response: Message + Default + 'static,

Execute the given requests a batch (same) request.

Note that this does not support the convenience types defined in the api module. You will need to convert them to the corresponding request type with Request::from.

pub fn auth_status(&self) -> AuthStatus

Get the current auth status.

Example

assert!(!client.auth_status().is_authenticated());

pub fn homeserver_url(&self) -> &Uri

Get the stored homeserver URL.

Example

assert_eq!(&client.homeserver_url().to_string(), "https://chat.harmonyapp.io:2289/");

pub fn make_hmc(&self, id: impl Display) -> Result<Hmc, HmcFromStrError>

Makes an HMC with homeserver’s authority and the given ID.

Example

assert_eq!(client.make_hmc("404").unwrap(), Hmc::new("chat.harmonyapp.io:2289", "404").unwrap());

pub fn begin_auth(
&self
) -> impl Future<Output = ClientResult<()>> + Send + 'static

Start an authentication session.

Example

client.begin_auth().await?;
// Do auth stuff here

pub fn next_auth_step(
&self,
response: AuthStepResponse
) -> impl Future<Output = ClientResult<Option<NextStepResponse>>> + Send + 'static

Request the next authentication step from the server.

Returns Ok(None) if authentication was completed. Returns Ok(Some(AuthStep)) if extra step is requested from the server.

Example

client.begin_auth().await?;
let next_step = client.next_auth_step(AuthStepResponse::Initial).await?;
// Do more auth stuff here

pub fn prev_auth_step(
&self
) -> impl Future<Output = ClientResult<StepBackResponse>> + Send + 'static

Go back to the previous authentication step.

Example

client.begin_auth().await?;
// Call next step and whatnot here
// Oops, user wants to do something else, lets go back
let prev_step = client.prev_auth_step().await?;
// Do more auth stuff here

pub fn auth_stream(
&self
) -> impl Future<Output = ClientResult<AuthSocket>> + Send + 'static

Begin an authentication steps stream for the current authentication session.

Example

client.begin_auth().await?;
let auth_steps_stream = client.auth_stream().await?;
// Do auth stuff here

pub fn subscribe_events(
&self,
subscriptions: Vec<EventSource>
) -> impl Future<Output = ClientResult<EventsSocket>> + Send + 'static

Subscribe to events coming from specified event sources.

Example

// Auth here
let event_stream = client.subscribe_events(vec![EventSource::Homeserver, EventSource::Action]).await?;
// Do more auth stuff here